[RGTC-565] Fix map for Internet Explorer 11
