+++ modules/custom/openy_map/js/map.js
@@ -1,1639 +1,1 @@
-/**
- * @file main.js
- */
-(function ($, window, Drupal, drupalSettings) {
-
-  "use strict";
-
-  Drupal.openyMap = function () {
-    return {
-      // Array of location data.
-      locations: null,
-      // URL of marker image.
-      marker_image_url: null,
-      // URL of shadow image.
-      shadow_image_url: null,
-      // The Map object.
-      map: null,
-      // Array of tag data, keyed by tag name.
-      tags: {},
-      // The distance filter limit, in miles.
-      distance_limit: null,
-      // The center point of the map.
-      center_point: null,
-      // The center point of a search location or distance limit.
-      search_center_point: null,
-      // Marker designating the center point.
-      search_center_marker: null,
-      // Geocoder.
-      geocoder: function() {
-        return typeof google.maps !== 'undefined' ? new google.maps.Geocoder() : {};
-      },
-
-      // Checks if the provider library object has loaded
-      libraryIsLoaded: function () {
-        return typeof window.google != 'undefined';
-      },
-
-      // Normalizes a map-vendor specific representation of
-      // a coordinate point to a {lat:x, lon:y} object.
-      normalize_point: function (point) {
-        return {
-          'lat': point.lat(),
-          'lon': point.lng()
-        };
-      },
-
-      // Convert a number from degrees to radians.
-      toRad: function (n) {
-        return n * Math.PI / 180;
-      },
-
-      init: function (args) {
-        this.component_el = args.component_el;
-
-        this.map_data = args.map_data;
-        this.tags_style = args.tags_style;
-        this.locations = this.map_data;
-
-        this.marker_image_url = args.marker_image_url || null;
-        this.shadow_image_url = args.shadow_image_url || null;
-
-        this.search_center_marker = args.search_center_marker || null;
-
-        this.map_el = this.component_el.find('.openy-map');
-        this.messages_el = this.component_el.find('.messages');
-
-        this.map_controls_el = this.component_el.find('.map_controls');
-        this.search_field_el = this.map_controls_el.find('input.search_field');
-        this.distance_limit_el = this.map_controls_el.find('select.distance_limit_value');
-        this.locate_me_el = this.map_controls_el.find('.locateme');
-
-        this.tags = {};
-        this.default_tags = drupalSettings.openyMapSettings.default_tags;
-        this.init_map();
-        this.init_tags();
-        this.init_map_locations();
-        this.draw_map_controls();
-        this.hookup_map_controls_events();
-        this.update_tag_filters();
-        this.update_amenities_filters();
-        this.draw_map_locations();
-        this.draw_list_locations();
-
-
-        var mapLocation = document.location.href.match(/&?[amp;]?map_location=([\w|\+]*)&?[amp;]?/),
-          component = this;
-
-        if (!navigator.geolocation) {
-          $('.with-geo').remove();
-        }
-        this.component_el.find('.zip-code .btn-submit')
-        .on('click', $.proxy(this.apply_search, this));
-
-        this.search_field_el.on('keypress', function (e) {
-          if (e.keyCode == 13) component.apply_search();
-        });
-        if (mapLocation) {
-          $('.search_field')
-          .val(mapLocation[1].replace(/\+/g, ' '));
-
-          $('.distance_limit option').eq(2).attr('selected', true);
-          $('.zip-code .btn-submit').click();
-        }
-      },
-
-      // Initializes the base map.
-      init_map: function () {
-        this.map = new google.maps.Map(this.map_el[0], {
-          scaleControl: true,
-          center: this.center,
-          zoom: 9,
-          scrollwheel: false,
-          mapTypeId: google.maps.MapTypeId.ROADMAP
-        });
-
-        this.init_map_center();
-
-        this.component_el.trigger('initialized', [this.map]);
-      },
-
-      init_map_center: function () {
-        this.search_center_marker = this.search_center_marker || new google.maps.Marker({
-            position: this.center_point,
-            animation: google.maps.Animation.DROP
-          });
-
-        if (this.search_center_marker) {
-          this.search_center_marker.setVisible(false);
-          this.search_center_marker.setMap(this.map);
-        }
-      },
-
-      // Executed every time a checkbox filter state changes.
-      filter_change: function (evt) {
-        if (evt) {
-          var evt_target = $(evt.currentTarget);
-          var matching_bar_checkbox = this.component_el.find('nav.types input.' + evt_target.attr('class') + '[type=checkbox]');
-          matching_bar_checkbox.prop('checked', evt_target.prop('checked'));
-          matching_bar_checkbox.parents('label').toggleClass('checked', evt_target.prop('checked'));
-        }
-
-        this.update_tag_filters();
-        this.update_amenities_filters();
-        this.redraw_map_locations();
-        this.draw_list_locations();
-      },
-
-      // Attaches events to various map controls.
-      hookup_map_controls_events: function () {
-        this.map_controls_el.find('.tag_filters input[type=checkbox]').on('change', $.proxy(this.filter_change, this));
-        this.search_field_el.on('change', $.proxy(this.apply_search, this));
-        this.distance_limit_el.on('change', $.proxy(this.apply_distance_limit, this));
-        this.locate_me_el.on('click', $.proxy(this.locate_me_onclick, this));
-        this.component_el.find('nav.types input[type=checkbox]').on('change', $.proxy(this.bar_filter_change, this));
-        this.search_field_el.on("autocompleteselect", $.proxy(this.apply_autocomplete_search, this));
-
-        $('#views-exposed-form-location-by-amenities-block-1').find('input[type=checkbox]').on('change', $.proxy(this.filter_change, this));
-      },
-
-      // Attempts a map search against Google's
-      // GeoCoding API.  If successful, the map
-      // is recentered according to the result.
-      apply_search: function () {
-        var q = this.search_field_el.val();
-        if (q == '') {
-          this.reset_search_results();
-          return;
-        }
-        var f = function (results, status) {
-          if (status == 'OK') {
-            this.search_center_point = results[0].geometry.location;
-
-            if (results[0].geometry.bounds) {
-              this.map.fitBounds(results[0].geometry.bounds);
-            } else {
-              var bounds = new google.maps.LatLngBounds();
-              bounds.extend(this.search_center_point);
-              // Don't zoom in too far on only one marker
-              if (bounds.getNorthEast().equals(bounds.getSouthWest())) {
-                var extendPoint1 = new google.maps.LatLng(bounds.getNorthEast().lat() + 0.001, bounds.getNorthEast().lng() + 0.001);
-                var extendPoint2 = new google.maps.LatLng(bounds.getNorthEast().lat() - 0.001, bounds.getNorthEast().lng() - 0.001);
-                bounds.extend(extendPoint1);
-                bounds.extend(extendPoint2);
-              }
-              this.map.fitBounds(bounds);
-            }
-
-            this.search_center = this.map.getCenter();
-            this.draw_search_center();
-            this.apply_distance_limit();
-          }
-        };
-
-        this.geocoder().geocode({
-          'address': q
-        }, $.proxy(f, this));
-      },
-
-      apply_autocomplete_search: function (event, ui) {
-        var locations = [];
-        this.locations.forEach(function (location) {
-          if (location.name == ui.item.value) {
-            // Get selected location from locations list.
-            locations.push(location);
-          }
-        });
-
-        // Redraw map for selected location.
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = '';
-        this.search_center_marker.setPosition(this.search_center_point);
-        this.search_center_marker.setVisible(false);
-        var bounds = new google.maps.LatLngBounds();
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          bounds.extend(loc.marker.getPosition());
-          loc.marker.setVisible(true);
-        }
-        this.map.fitBounds(bounds);
-
-        // Redraw locations list.
-        for (var l = 0; l < this.locations.length; l++) {
-          if (typeof this.locations[l].element !== 'undefined') {
-            this.locations[l].element.hide();
-            $(this.locations[l].element).parents('.locations-list').find('.location-title').hide();
-          }
-        }
-
-        if (!locations.length) {
-          var message_html = '<div class="col-xs-12 text-center"><p>' +
-            Drupal.t("We're sorry no results were found in your area") +
-            '</p></div>';
-          this.messages_el.hide().html(message_html).fadeIn();
-          return;
-        }
-        // Show filtered locations.
-        for (var k = 0; k < locations.length; k++) {
-          if (typeof locations[k].element !== 'undefined') {
-            locations[k].element.show();
-            $(locations[k].element).parents('.locations-list').find('.location-title').show();
-          }
-        }
-      },
-
-      // Executed every time the viewer sets the distance limit to a new value.
-      apply_distance_limit: function () {
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = this.distance_limit_el.val();
-
-        this.draw_search_center();
-        this.redraw_map_locations();
-        this.draw_list_locations();
-      },
-
-
-      // Executed if was provided empty ZIP code.
-      reset_search_results: function () {
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = '';
-        this.search_center_marker.setPosition(this.search_center_point);
-        this.search_center_marker.setVisible(false);
-        this.redraw_map_locations();
-        this.draw_list_locations();
-      },
-
-      locate_me_onclick: function (evt) {
-        if (!navigator.geolocation) {
-          return;
-        }
-
-        this.search_field_el.val('');
-        this.geolocation_watcher = navigator.geolocation.watchPosition($.proxy(this.locate_me, this));
-      },
-
-      locate_me: function (position) {
-        var lat = position.coords.lat;
-        var lng = position.coords.lng;
-
-        this.search_center_point = new google.maps.LatLng(lat, lng);
-
-        this.map.setCenter(this.search_center_point);
-        this.map.setZoom(14);
-        if (position.coords.accuracy <= 15840) { // 3 miles.
-
-          this.geocoder().geocode({
-              'latLng': this.search_center_point
-            },
-            $.proxy(
-              function (results, status) {
-                if (results[0]) this.search_field_el.val(results[0].formatted_address);
-                this.apply_search();
-              },
-              this));
-
-          navigator.geolocation.clearWatch(this.geolocation_watcher);
-        }
-
-        this.draw_search_center();
-      },
-
-      // Extracts unique tag values from the map location data.
-      init_tags: function () {
-        // Extract tags.
-        for (var i = 0; i < this.locations.length; i++) {
-          var loc = this.locations[i];
-          if (!loc.tags) {
-            loc.tags = [];
-          }
-
-          // Convert single-string tags to array.
-          if (typeof(loc.tags) == typeof( "" )) {
-            loc.tags = [loc.tags];
-          }
-
-          for (var j = 0; j < loc.tags.length; j++) {
-            var tag = loc.tags[j];
-            if (!( tag in this.tags )) {
-              this.tags[tag] = {'marker_icons': []};
-            }
-            if (loc.icon && $.inArray(loc.icon, this.tags[tag].marker_icons) == -1) {
-              this.tags[tag].marker_icons.push(loc.icon);
-            }
-          }
-        }
-      },
-
-
-      // Applies the current checkbox state of the tag filter controls
-      // to the internal filters data structure.
-      // Called at init time, and after every checkbox state change.
-      update_tag_filters: function () {
-        this.tag_filters = [];
-        var self = this;
-
-        var f = function (index) {
-          var el = $(this);
-          self.tag_filters.push(el.val());
-        };
-        this.map_controls_el.find('.tag_filters input[type=checkbox]:checked').each(f);
-      },
-
-      update_amenities_filters: function () {
-        this.amenities_filters = [];
-        var self = this;
-
-        var f = function (index) {
-          var el = $(this);
-          self.amenities_filters.push(el.val());
-        };
-
-        $('#views-exposed-form-location-by-amenities-block-1')
-        .find('input[type=checkbox]:checked')
-        .each(f);
-      },
-
-      // Applies tag and distance filters to a list of locations,
-      // returns the filtered list.
-      apply_filters: function (locations) {
-        locations = this.apply_tag_filters(locations);
-        locations = this.apply_distance_filters(locations);
-        locations = this.apply_amenities_filters(locations);
-        this.set_url_parameters();
-        return locations;
-      },
-
-      // Applies tag filters to a list of locations,
-      // returns the filtered list.
-      apply_tag_filters: function (locations) {
-        var selected_tags_count = this.tag_filters.length;
-        var tags_count = Object.keys(this.tags).length;
-        var show_facilities = tags_count > this.default_tags.length;
-        if (selected_tags_count === 0 || (selected_tags_count === tags_count && show_facilities)) {
-          // Return all locations if:
-          // - Tags not selected.
-          // - Selected all tags and we have more items than implemented
-          //   in default_tags(exist at least one facility type).
-          return locations;
-        }
-
-        var filtered_locations = [];
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          for (var j = 0; j < this.tag_filters.length; j++) {
-            var tag_filter = this.tag_filters[j];
-            if ($.inArray(tag_filter, loc.tags) >= 0) {
-              filtered_locations.push(loc);
-              continue;  // If any tag matches, skip checking other tags.
-            }
-          }
-        }
-        return filtered_locations;
-      },
-
-      // Applies distance filters to a list of locations,
-      // returns the filtered list.
-      apply_distance_filters: function (locations) {
-        if (!this.search_center) {
-          return locations;
-        }
-
-        if (!this.distance_limit || this.distance_limit === '') {
-          return locations;
-        }
-
-        var search_center = this.normalize_point(this.search_center);
-        var filtered_locations = [];
-
-        var lat1 = parseFloat(search_center.lat);
-        var lon1 = parseFloat(search_center.lon);
-        var rlat1 = this.toRad(lat1);
-
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          var R = 3963,
-            lat2 = parseFloat(loc.lat),
-            lon2 = parseFloat(loc.lng);
-
-          var rlat = this.toRad(lat2 - lat1);
-          var rlon = this.toRad(lon2 - lon1);
-          var rlat2 = this.toRad(lat2);
-
-          var a = Math.sin(rlat / 2) * Math.sin(rlat / 2) + Math.sin(rlon / 2) * Math.sin(rlon / 2) * Math.cos(rlat1) * Math.cos(rlat2);
-          var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-          var d = R * c;
-
-          if (d <= this.distance_limit) {
-            // Add the distance to the object.
-            loc.distance = d;
-            filtered_locations.push(loc);
-          }
-        }
-
-        return filtered_locations;
-      },
-
-
-      // Applies tag filters to a list of locations,
-      // returns the filtered list.
-      apply_amenities_filters: function (locations) {
-        if (this.amenities_filters.length === 0) {
-          return locations;
-        }
-        var filtered_locations = [];
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          for (var j = 0; j < this.amenities_filters.length; j++) {
-            var amenities_filter = this.amenities_filters[j];
-            if ($.inArray(amenities_filter, loc.amenities) >= 0) {
-              filtered_locations.push(loc);
-              continue;  // If any tag matches, skip checking other tags
-            }
-          }
-        }
-        return filtered_locations;
-      },
-
-
-      // Populates an array of active tags from an URL parameter "type".
-      init_active_tags: function () {
-        if (this.initial_active_tags) {
-          return this.initial_active_tags;
-        }
-
-        var active_tags = [];
-        var url_parameters = this.get_parameters();
-        var tag_filter_url_value = url_parameters.type;
-
-        var tag_filter_url_values = ( tag_filter_url_value ) ? tag_filter_url_value.split(",") : [];
-
-        for (var tag in this.tags) {
-          if (tag_filter_url_values.length === 0) {
-            active_tags.push(tag);
-          }
-          else if ($.inArray(this.encode_to_url_format(tag), tag_filter_url_values) >= 0) {
-            active_tags.push(tag);
-          }
-        }
-
-        this.initial_active_tags = active_tags;
-
-        if (tag_filter_url_values.length === 0) {
-          this.initial_active_tags = this.default_tags;
-        }
-      },
-
-      // Get url params.
-      get_parameters: function () {
-        var searchString = window.location.search.substring(1);
-        var params = searchString.split("&");
-        var hash = {};
-
-        for (var i = 0; i < params.length; i++) {
-          var val = params[i].split("=");
-          hash[unescape(val[0])] = unescape(val[1]);
-        }
-        return hash;
-      },
-
-      // Update url params.
-      set_url_parameters: function () {
-        var url = document.location.pathname,
-          params = this.get_parameters(),
-          filterTagsRaw = this.tag_filters,
-          filteramenitiesRaw = this.amenities_filters,
-          filterTags = '',
-          filteramenities = '',
-          mapLocation = $('.search_field').val() || (params.hasOwnProperty('map_location') && params.map_location) || '';
-        if (mapLocation) {
-          mapLocation = '?map_location=' + this.encode_to_url_format(mapLocation);
-        }
-        if (filterTagsRaw) {
-          filterTags = !mapLocation ? '?' : '&';
-          filterTags += 'type=';
-          filterTagsRaw.forEach(function (tag) {
-            filterTags += this.encode_to_url_format(tag) + ',';
-          }, this, filterTags);
-          filterTags = filterTags.substring(0, filterTags.length - 1);
-        }
-        if (filteramenitiesRaw) {
-          filteramenities = '&';
-          filteramenities += 'amenities=';
-          filteramenitiesRaw.forEach(function (tag) {
-            filteramenities += this.encode_to_url_format(tag) + ',';
-          }, this, filteramenities);
-          filteramenities = filteramenities.substring(0, filteramenities.length - 1);
-        }
-        window.history.replaceState(null, null, url + mapLocation + filterTags + filteramenities);
-      },
-
-      // Renders an extra set of filter boxes below the map.
-      draw_map_controls: function () {
-        // Add tag filter to map control.
-        this.init_active_tags();
-        var tag_filters_html = '';
-
-        if (this.tags_style == 'list-checkboxes') {
-          // Show tags filter as multiselect list checkboxes.
-          tag_filters_html = '<select id="tag-filter" class="form-control multiselect" name="tag_filter" multiple="multiple">';
-          // Sort tags alphabetically.
-          var tags = Object.keys(this.tags).sort();
-          // Move YMCA and Camps tags to begin.
-          // @todo Names are hardcoded, this should be refactored to allow use any name we want.
-          tags.splice(tags.indexOf('YMCA'), 1);
-          tags.splice(tags.indexOf('Camps'), 1);
-          tags.unshift('Camps');
-          tags.unshift('YMCA');
-          tags.forEach(function (tag) {
-            var filter_checked = '';
-            if ($.inArray(tag, this.initial_active_tags) >= 0) {
-              filter_checked = 'selected';
-            }
-            tag_filters_html += '<option value="' + tag + '" ' + filter_checked + '>' + tag + '</option>';
-          }, this);
-          tag_filters_html += '</select>';
-        }
-        else {
-          // Show tags filter as default checkboxes.
-          for (var tag in this.tags) {
-            var filter_checked = '';
-            if ($.inArray(tag, this.initial_active_tags) >= 0) {
-              filter_checked = 'checked="checked"';
-            }
-            var tag_filter_html = '<label class="btn btn-default" for="tag_' + tag + '">';
-            tag_filter_html += '<input autocomplete="off" id="tag_' + tag + '" class="tag_' + tag + '" type="checkbox" value="' + tag + '" ' + filter_checked + '/>' + tag;
-            for (var i = 0; i < this.tags[tag].marker_icons.length; i++) {
-              tag_filter_html += '<img class="tag_icon inline-hidden-sm" src="' + this.tags[tag].marker_icons[i] + '" aria-hidden="true" />';
-            }
-            tag_filter_html += '</label>';
-            tag_filters_html += tag_filter_html;
-          }
-        }
-
-        this.map_controls_el.find('.tag_filters').append(tag_filters_html);
-
-        if (this.tags_style == 'list-checkboxes') {
-          // Init multiselect if used list-checkboxes.
-          $(".tag_filters .multiselect").multiselect({
-            columns: 1,
-            showCheckbox: true,
-            minHeight: 50,
-            texts: {
-              placeholder: 'Select options'
-            },
-            onOptionClick: $.proxy(this.filter_change, this)
-          });
-        }
-        // Add locations autocomplete to search field.
-        var locations = [];
-        this.locations.forEach(function (location) {
-          locations.push(location.name);
-        });
-        this.search_field_el.autocomplete({
-          minLength: 3,
-          source: locations
-        });
-      },
-
-      // Convert string to url format:
-      // remove all non-alphanumeric characters, convert to lowercase,
-      // replace spaces with dashes.
-      encode_to_url_format: function (txt) {
-        return txt
-        .toLowerCase()
-        .replace(/[^\w ]+/g, '')
-        .replace(/ +/g, '-')
-          ;
-      },
-
-      // Update locations on the map by setting their visibility
-      // and refit the map bounds to the current set of visible locations.
-      draw_map_locations: function () {
-        var locations = this.apply_filters(this.locations);
-
-        // If the location list is empty, don't adjust the map at all.
-        if (locations.length === 0) {
-          this.map.setCenter(this.search_center_point);
-          return;
-        }
-
-        var bounds = new google.maps.LatLngBounds();
-
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          bounds.extend(loc.marker.getPosition());
-          loc.marker.setVisible(true);
-        }
-
-        // Don't zoom in too far on only one marker.
-        if (bounds.getNorthEast().equals(bounds.getSouthWest())) {
-          var extendPoint1 = new google.maps.LatLng(bounds.getNorthEast().lat() + 0.001, bounds.getNorthEast().lng() + 0.001);
-          var extendPoint2 = new google.maps.LatLng(bounds.getNorthEast().lat() - 0.001, bounds.getNorthEast().lng() - 0.001);
-          bounds.extend(extendPoint1);
-          bounds.extend(extendPoint2);
-        }
-        this.map.fitBounds(bounds);
-
-      },
-
-      // Updates locations on the map by setting their visibility
-      // to false before drawing.
-      redraw_map_locations: function () {
-        for (var i = 0; i < this.locations.length; i++) {
-          var loc = this.locations[i];
-          loc.marker.setVisible(false);
-        }
-
-        this.draw_map_locations();
-      },
-
-      // Render the list of locations.
-      draw_list_locations: function () {
-
-        var locations = this.apply_filters(this.locations);
-
-        // Hide all heading locations.
-        for (var l = 0; l < this.locations.length; l++) {
-          if (typeof this.locations[l].element !== 'undefined') {
-            this.locations[l].element.hide();
-            $(this.locations[l].element).parents('.locations-list').find('.location-title').hide();
-          }
-        }
-
-        if (!locations.length) {
-          var message_html = '<div class="col-xs-12 text-center"><p>' +
-            Drupal.t('No locations were found in this area. Please try a different area or increase your search distance.') +
-            '</p></div>';
-          this.messages_el.hide().html(message_html).fadeIn();
-          return;
-        }
-        else {
-          this.messages_el.hide();
-        }
-
-        // Show filtered locations.
-        for (var k = 0; k < locations.length; k++) {
-          if (typeof locations[k].element !== 'undefined') {
-            locations[k].element.show();
-            $(locations[k].element).parents('.locations-list').find('.location-title').show();
-          }
-        }
-      },
-
-      // Generate the HTML for a single location in the list.
-      draw_list_location: function (loc) {
-        return loc.markup;
-      },
-
-      // Generate the HTML for a single location's map detail view
-      draw_map_location: function (loc) {
-        return this.draw_list_location(loc);
-      },
-
-      init_map_locations: function () {
-        var self = this;
-        var locations = this.locations;
-
-        var f = function (infowindow, marker) {
-          return function () {
-            for (var i = 0; i < self.locations.length; i++) {
-              self.locations[i].infowindow.close();
-            }
-            infowindow.open(this.map, marker);
-          };
-        };
-
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          loc.point = new google.maps.LatLng(loc.lat, loc.lng);
-          var html = '<div class="marker_tooltip">' + this.draw_map_location(loc) + '</div>';
-          var marker_anchor = new google.maps.MarkerImage(this.marker_image_url) || null;
-          marker_anchor = loc.icon ? new google.maps.MarkerImage(loc.icon) : marker_anchor;
-          var shadow_anchor = loc.shadow ? new google.maps.MarkerImage(loc.shadow) : null;
-
-          var marker = new google.maps.Marker({
-            position: loc.point,
-            icon: marker_anchor,
-            shadow: shadow_anchor,
-            animation: google.maps.Animation.DROP
-          });
-
-          var infowindow = new google.maps.InfoWindow({
-            content: html
-          });
-
-          loc.infowindow = infowindow;
-
-          google.maps.event.addListener(marker, 'click', f(infowindow, marker));
-
-          marker.setVisible(false);
-          marker.setMap(this.map);
-          loc.marker = marker;
-        }
-      },
-
-      draw_search_center: function () {
-        this.search_center_marker.setPosition(this.search_center_point);
-        this.search_center_marker.setVisible(true);
-      },
-
-      // Executed every time a checkbox bar filter state changes.
-      bar_filter_change: function (evt) {
-        var evt_target = $(evt.currentTarget);
-        var matching_checkbox = this.map_controls_el.find('input.' + evt_target.attr('class') + '[type=checkbox]');
-
-        // Uncheck all checkboxes.
-        this.map_controls_el.find('input[type=checkbox]').prop('checked', false);
-        this.component_el.find('nav.types input[type=checkbox]').prop('checked', false);
-        this.component_el.find('nav.types label').removeClass('checked');
-
-        // Check the just-clicked one.
-        matching_checkbox.prop('checked', true);
-        evt_target.prop('checked', true);
-
-        evt_target.parents('label').addClass('checked');
-        this.filter_change();
-      },
-
-      build_google_url: function (str) {
-        str = str.trim();
-        str = str.replace(/ /g, "+");
-
-        if (str.length > 0) {
-          str += '+';
-        }
-        return str;
-      }
-    };
-  };
-
-  Drupal.baseLayerWikimedia = {
-    tilePattern: 'https://maps.wikimedia.org/osm-intl/{z}/{x}/{y}{r}.png',
-    options: {
-      attribution: '<a href="https://wikimediafoundation.org/wiki/Maps_Terms_of_Use">Wikimedia</a>',
-      minZoom: 1,
-      maxZoom: 19
-    }
-  };
-
-  Drupal.baseLayerEsriWorldStreetMap = {
-    tilePattern: 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer/tile/{z}/{y}/{x}',
-    options: {
-      attribution: 'Tiles &copy; Esri &mdash; Source: Esri, DeLorme, NAVTEQ, USGS, Intermap, iPC, NRCAN, Esri Japan, METI, Esri China (Hong Kong), Esri (Thailand), TomTom, 2012'
-    }
-  };
-
-  Drupal.baseLayerEsriNatGeoWorldMap = {
-    tilePattern: 'https://server.arcgisonline.com/ArcGIS/rest/services/NatGeo_World_Map/MapServer/tile/{z}/{y}/{x}',
-    options: {
-      attribution: 'Tiles &copy; Esri &mdash; National Geographic, Esri, DeLorme, NAVTEQ, UNEP-WCMC, USGS, NASA, ESA, METI, NRCAN, GEBCO, NOAA, iPC',
-      maxZoom: 16
-    }
-  };
-
-  Drupal.baseLayerOpenStreetMapMapnik = {
-    tilePattern: 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-    options: {
-      attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>',
-      maxZoom: 19
-    }
-  };
-
-  Drupal.openyMapLeaflet = function () {
-    return {
-      baseLayer: Drupal.baseLayerWikimedia,
-      // Array of location data.
-      locations: null,
-      // URL of marker image.
-      marker_image_url: null,
-      // URL of shadow image.
-      shadow_image_url: null,
-      // The Map object.
-      map: null,
-      // Array of tag data, keyed by tag name.
-      tags: {},
-      // The distance filter limit, in miles.
-      distance_limit: null,
-      // The center point of the map.
-      center_point: null,
-      // The center point of a search location or distance limit.
-      search_center_point: null,
-      // Marker designating the center point.
-      search_center_marker: null,
-      // The paddings for fitBounds method, depends on marker dimensions.
-      fitBoundsOptions: null,
-      // Suffix for nominatim geocoder
-      default_search_location: null,
-
-      // Checks if the provider library object has loaded
-      libraryIsLoaded: function () {
-        return typeof window.L != 'undefined';
-      },
-
-      // Normalizes a map-vendor specific representation of
-      // a coordinate point to a {lat:x, lon:y} object.
-      normalize_point: function (point) {
-        return {
-          'lat': point.lat(),
-          'lon': point.lng()
-        };
-      },
-
-      // Convert a number from degrees to radians.
-      toRad: function (n) {
-        return n * Math.PI / 180;
-      },
-
-      init: function (args) {
-        this.component_el = args.component_el;
-
-        this.map_data = args.map_data;
-        this.tags_style = args.tags_style;
-        this.locations = this.map_data;
-
-        // Depends on markers' dimensions.
-        this.fitBoundsOptions = {
-          paddingTopLeft: L.point(0, 40),
-          paddingBottomRight: L.point(0, 10)
-        };
-
-        this.marker_image_url = args.marker_image_url || null;
-        this.shadow_image_url = args.shadow_image_url || null;
-
-        this.search_center_marker = args.search_center_marker || null;
-
-        this.map_el = this.component_el.find('.openy-map');
-        this.messages_el = this.component_el.find('.messages');
-
-        this.map_controls_el = this.component_el.find('.map_controls');
-        this.search_field_el = this.map_controls_el.find('input.search_field');
-        this.distance_limit_el = this.map_controls_el.find('select.distance_limit_value');
-        this.locate_me_el = this.map_controls_el.find('.locateme');
-
-        this.tags = {};
-        this.default_tags = drupalSettings.openyMapSettings.default_tags;
-        this.init_map();
-        this.init_tags();
-        this.init_map_locations();
-        this.draw_map_controls();
-        this.hookup_map_controls_events();
-        this.update_tag_filters();
-        this.update_amenities_filters();
-        this.draw_map_locations();
-        this.draw_list_locations();
-
-
-        var mapLocation = document.location.href.match(/&?[amp;]?map_location=([\w|\+]*)&?[amp;]?/),
-            component = this;
-
-        if (!navigator.geolocation) {
-          $('.with-geo').remove();
-        }
-        this.component_el.find('.zip-code .btn-submit')
-            .on('click', $.proxy(this.apply_search, this));
-
-        this.search_field_el.on('keypress', function (e) {
-          if (e.keyCode == 13) component.apply_search();
-        });
-        if (mapLocation) {
-          $('.search_field')
-              .val(mapLocation[1].replace(/\+/g, ' '));
-
-          $('.distance_limit option').eq(2).attr('selected', true);
-          $('.zip-code .btn-submit').click();
-        }
-      },
-
-      // Initializes the base map.
-      init_map: function () {
-        this.map = L.map(this.map_el[0]).setView([51.505, -0.09], 13);
-        L.tileLayer(this.baseLayer.tilePattern, this.baseLayer.options).addTo(this.map);
-        this.map.scrollWheelZoom.disable();
-        this.init_map_center();
-      },
-
-      init_map_center: function () {
-        var icon = L.icon({
-          iconUrl: this.search_icon,
-          iconRetinaUrl: this.search_icon_retina,
-          shadowUrl: 'https://cdnjs.cloudflare.com/ajax/libs/leaflet/0.7.7/images/marker-shadow.png',
-          iconSize: [25, 41],
-          iconAnchor: [12, 41],
-          popupAnchor: [1, -34],
-          shadowSize: [41, 41]
-        });
-        this.search_center_marker = this.search_center_marker || L.marker(this.map.getCenter(), {icon: icon});
-
-        if (this.search_center_marker) {
-          this.search_center_marker.removeFrom(this.map);
-        }
-      },
-
-      // Executed every time a checkbox filter state changes.
-      filter_change: function (evt) {
-        if (evt) {
-          var evt_target = $(evt.currentTarget);
-          var matching_bar_checkbox = this.component_el.find('nav.types input.' + evt_target.attr('class') + '[type=checkbox]');
-          matching_bar_checkbox.prop('checked', evt_target.prop('checked'));
-          matching_bar_checkbox.parents('label').toggleClass('checked', evt_target.prop('checked'));
-        }
-
-        this.update_tag_filters();
-        this.update_amenities_filters();
-        this.redraw_map_locations();
-        this.draw_list_locations();
-      },
-
-      // Attaches events to various map controls.
-      hookup_map_controls_events: function () {
-        this.map_controls_el.find('.tag_filters input[type=checkbox]').on('change', $.proxy(this.filter_change, this));
-        this.search_field_el.on('change', $.proxy(this.apply_search, this));
-        this.distance_limit_el.on('change', $.proxy(this.apply_distance_limit, this));
-        this.locate_me_el.on('click', $.proxy(this.locate_me_onclick, this));
-        this.component_el.find('nav.types input[type=checkbox]').on('change', $.proxy(this.bar_filter_change, this));
-        this.search_field_el.on("autocompleteselect", $.proxy(this.apply_autocomplete_search, this));
-
-        $('#views-exposed-form-location-by-amenities-block-1').find('input[type=checkbox]').on('change', $.proxy(this.filter_change, this));
-      },
-
-      // Attempts a map search against OSM Nominatim API. If successful, the map
-      // is recentered according to the result.
-      apply_search: function () {
-        var self = this;
-        var q = this.search_field_el.val();
-        if (q == '') {
-          this.reset_search_results();
-          return;
-        }
-
-        this.geocode(q, function(data, status) {
-          if (status == 'success' && data.length > 0) {
-            self.search_center_point = L.latLng(data[0].lat, data[0].lon);
-
-            if (data[0].boundingbox) {
-              var bounds = L.latLngBounds();
-              bounds.extend(L.latLng(data[0].boundingbox[0], data[0].boundingbox[2]));
-              bounds.extend(L.latLng(data[0].boundingbox[1], data[0].boundingbox[3]));
-              self.map.fitBounds(bounds, self.fitBoundsOptions);
-            }
-
-            self.search_center = self.search_center_point;
-            self.draw_search_center();
-            self.apply_distance_limit();
-          }
-        });
-      },
-
-      geocode: function(query, callback) {
-        var base = 'https://nominatim.openstreetmap.org/search?format=json&limit=5&q=';
-        var suffix = this.default_search_location ? '+' + this.default_search_location : '';
-        $.getJSON(base + query + suffix, callback);
-      },
-
-      apply_autocomplete_search: function (event, ui) {
-        var locations = [];
-        this.locations.forEach(function (location) {
-          if (location.name == ui.item.value) {
-            // Get selected location from locations list.
-            locations.push(location);
-          }
-        });
-
-        // Redraw map for selected location.
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = '';
-        this.search_center_marker.removeFrom(this.maps);
-        var bounds = L.latLngBounds();
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          bounds.extend(loc.point);
-          loc.marker.addTo(this.map);
-        }
-        this.map.fitBounds(bounds, this.fitBoundsOptions);
-
-        // Redraw locations list.
-        for (var l = 0; l < this.locations.length; l++) {
-          if (typeof this.locations[l].element !== 'undefined') {
-            this.locations[l].element.hide();
-            $(this.locations[l].element).parents('.locations-list').find('.location-title').hide();
-          }
-        }
-
-        if (!locations.length) {
-          var message_html = '<div class="col-xs-12 text-center"><p>' +
-              Drupal.t("We're sorry no results were found in your area") +
-              '</p></div>';
-          this.messages_el.hide().html(message_html).fadeIn();
-          return;
-        }
-        // Show filtered locations.
-        for (var k = 0; k < locations.length; k++) {
-          if (typeof locations[k].element !== 'undefined') {
-            locations[k].element.show();
-            $(locations[k].element).parents('.locations-list').find('.location-title').show();
-          }
-        }
-      },
-
-      // Executed every time the viewer sets the distance limit to a new value.
-      apply_distance_limit: function () {
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = this.distance_limit_el.val();
-
-        this.draw_search_center();
-        this.redraw_map_locations();
-        this.draw_list_locations();
-      },
-
-
-      // Executed if was provided empty ZIP code.
-      reset_search_results: function () {
-        if (this.search_center === null) {
-          this.search_center = this.map.getCenter();
-        }
-        this.distance_limit = '';
-        if (this.search_center_point) {
-          this.search_center_marker.setLatLng(this.search_center_point);
-          this.search_center_marker.addTo(this.map);
-        }
-        this.redraw_map_locations();
-        this.draw_list_locations();
-        this.set_url_parameters();
-      },
-
-      locate_me_onclick: function (evt) {
-        if (!navigator.geolocation) {
-          return;
-        }
-
-        this.search_field_el.val('');
-        this.geolocation_watcher = navigator.geolocation.watchPosition($.proxy(this.locate_me, this));
-      },
-
-      locate_me: function (position) {
-        var lat = position.coords.lat;
-        var lng = position.coords.lng;
-
-        this.search_center_point = L.latLng(lat, lng);
-
-        this.map.setView(this.search_center_point, 14);
-        if (position.coords.accuracy <= 15840) { // 3 miles.
-          navigator.geolocation.clearWatch(this.geolocation_watcher);
-        }
-
-        this.draw_search_center();
-      },
-
-      // Extracts unique tag values from the map location data.
-      init_tags: function () {
-        // Extract tags.
-        for (var i = 0; i < this.locations.length; i++) {
-          var loc = this.locations[i];
-          if (!loc.tags) {
-            loc.tags = [];
-          }
-
-          // Convert single-string tags to array.
-          if (typeof(loc.tags) == typeof( "" )) {
-            loc.tags = [loc.tags];
-          }
-
-          for (var j = 0; j < loc.tags.length; j++) {
-            var tag = loc.tags[j];
-            if (!( tag in this.tags )) {
-              this.tags[tag] = {'marker_icons': []};
-            }
-            if (loc.icon && $.inArray(loc.icon, this.tags[tag].marker_icons) == -1) {
-              this.tags[tag].marker_icons.push(loc.icon);
-            }
-          }
-        }
-      },
-
-      // Applies the current checkbox state of the tag filter controls
-      // to the internal filters data structure.
-      // Called at init time, and after every checkbox state change.
-      update_tag_filters: function () {
-        this.tag_filters = [];
-        var self = this;
-
-        var f = function (index) {
-          var el = $(this);
-          self.tag_filters.push(el.val());
-        };
-        this.map_controls_el.find('.tag_filters input[type=checkbox]:checked').each(f);
-      },
-
-      update_amenities_filters: function () {
-        this.amenities_filters = [];
-        var self = this;
-
-        var f = function (index) {
-          var el = $(this);
-          self.amenities_filters.push(el.val());
-        };
-
-        $('#views-exposed-form-location-by-amenities-block-1')
-            .find('input[type=checkbox]:checked')
-            .each(f);
-      },
-
-      // Applies tag and distance filters to a list of locations,
-      // returns the filtered list.
-      apply_filters: function (locations) {
-        locations = this.apply_tag_filters(locations);
-        locations = this.apply_distance_filters(locations);
-        locations = this.apply_amenities_filters(locations);
-        this.set_url_parameters();
-        return locations;
-      },
-
-      // Applies tag filters to a list of locations,
-      // returns the filtered list.
-      apply_tag_filters: function (locations) {
-        var selected_tags_count = this.tag_filters.length;
-        var tags_count = Object.keys(this.tags).length;
-        var show_facilities = tags_count > this.default_tags.length;
-        if (selected_tags_count === 0 || (selected_tags_count === tags_count && show_facilities)) {
-          // Return all locations if:
-          // - Tags not selected.
-          // - Selected all tags and we have more items than implemented
-          //   in default_tags(exist at least one facility type).
-          return locations;
-        }
-
-        var filtered_locations = [];
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          for (var j = 0; j < this.tag_filters.length; j++) {
-            var tag_filter = this.tag_filters[j];
-            if ($.inArray(tag_filter, loc.tags) >= 0) {
-              filtered_locations.push(loc);
-              continue;  // If any tag matches, skip checking other tags.
-            }
-          }
-        }
-        return filtered_locations;
-      },
-
-      // Applies distance filters to a list of locations,
-      // returns the filtered list.
-      apply_distance_filters: function (locations) {
-        if (!this.search_center) {
-          return locations;
-        }
-
-        if (!this.distance_limit || this.distance_limit === '') {
-          return locations;
-        }
-
-        var search_center = this.search_center;
-        var filtered_locations = [];
-
-        var lat1 = parseFloat(search_center.lat);
-        var lon1 = parseFloat(search_center.lng);
-        var rlat1 = this.toRad(lat1);
-
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          var R = 3963,
-              lat2 = parseFloat(loc.point.lat),
-              lon2 = parseFloat(loc.point.lng);
-
-          var rlat = this.toRad(lat2 - lat1);
-          var rlon = this.toRad(lon2 - lon1);
-          var rlat2 = this.toRad(lat2);
-
-          var a = Math.sin(rlat / 2) * Math.sin(rlat / 2) + Math.sin(rlon / 2) * Math.sin(rlon / 2) * Math.cos(rlat1) * Math.cos(rlat2);
-          var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-          var d = R * c;
-
-          if (d <= this.distance_limit) {
-            // Add the distance to the object.
-            loc.distance = d;
-            filtered_locations.push(loc);
-          }
-        }
-
-        return filtered_locations;
-      },
-
-
-      // Applies tag filters to a list of locations,
-      // returns the filtered list.
-      apply_amenities_filters: function (locations) {
-        if (this.amenities_filters.length === 0) {
-          return locations;
-        }
-        var filtered_locations = [];
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          for (var j = 0; j < this.amenities_filters.length; j++) {
-            var amenities_filter = this.amenities_filters[j];
-            if ($.inArray(amenities_filter, loc.amenities) >= 0) {
-              filtered_locations.push(loc);
-              continue;  // If any tag matches, skip checking other tags
-            }
-          }
-        }
-        return filtered_locations;
-      },
-
-
-      // Populates an array of active tags from an URL parameter "type".
-      init_active_tags: function () {
-        if (this.initial_active_tags) {
-          return this.initial_active_tags;
-        }
-
-        var active_tags = [];
-        var url_parameters = this.get_parameters();
-        var tag_filter_url_value = url_parameters.type;
-
-        var tag_filter_url_values = ( tag_filter_url_value ) ? tag_filter_url_value.split(",") : [];
-
-        for (var tag in this.tags) {
-          if (tag_filter_url_values.length === 0) {
-            active_tags.push(tag);
-          }
-          else if ($.inArray(this.encode_to_url_format(tag), tag_filter_url_values) >= 0) {
-            active_tags.push(tag);
-          }
-        }
-
-        this.initial_active_tags = active_tags;
-
-        if (tag_filter_url_values.length === 0) {
-          this.initial_active_tags = this.default_tags;
-        }
-      },
-
-      // Get url params.
-      get_parameters: function () {
-        var searchString = window.location.search.substring(1);
-        var params = searchString.split("&");
-        var hash = [];
-
-        for (var i = 0; i < params.length; i++) {
-          var val = params[i].split("=");
-          hash[unescape(val[0])] = unescape(val[1]);
-        }
-        return hash;
-      },
-
-      // Update url params.
-      set_url_parameters: function () {
-        var url = document.location.pathname,
-            params = this.get_parameters(),
-            filterTagsRaw = this.tag_filters,
-            filteramenitiesRaw = this.amenities_filters,
-            filterTags = '',
-            filteramenities = '',
-            utmListRaw = {},
-            utms = '',
-            mapLocation = $('.search_field').val() || (params.hasOwnProperty('map_location') && params.map_location) || '';
-
-        for (let [key, value] of Object.entries(params)) {
-          var prmsKey = key.split("_");
-          if (prmsKey[0] === 'utm'){
-            utmListRaw[key] = value;
-          }
-        }
-
-        if (Object.keys(utmListRaw).length) {
-          utms = '&';
-          utms += jQuery.param(utmListRaw);
-        }
-
-        if (mapLocation) {
-          mapLocation = '?map_location=' + this.encode_to_url_format(mapLocation);
-        }
-        if (filterTagsRaw) {
-          filterTags = !mapLocation ? '?' : '&';
-          filterTags += 'type=';
-          filterTagsRaw.forEach(function (tag) {
-            filterTags += this.encode_to_url_format(tag) + ',';
-          }, this, filterTags);
-          filterTags = filterTags.substring(0, filterTags.length - 1);
-        }
-        if (filteramenitiesRaw) {
-          filteramenities = '&';
-          filteramenities += 'amenities=';
-          filteramenitiesRaw.forEach(function (tag) {
-            filteramenities += this.encode_to_url_format(tag) + ',';
-          }, this, filteramenities);
-          filteramenities = filteramenities.substring(0, filteramenities.length - 1);
-        }
-        window.history.replaceState(null, null, url + mapLocation + filterTags + filteramenities + utms);
-      },
-
-      // Renders an extra set of filter boxes below the map.
-      draw_map_controls: function () {
-        // Add tag filter to map control.
-        this.init_active_tags();
-        var tag_filters_html = '';
-
-        if (this.tags_style == 'list-checkboxes') {
-          // Show tags filter as multiselect list checkboxes.
-          tag_filters_html = '<select id="tag-filter" class="form-control multiselect" name="tag_filter" multiple="multiple">';
-          // Sort tags alphabetically.
-          var tags = Object.keys(this.tags).sort();
-          // Move YMCA and Camps tags to begin.
-          // @todo Names are hardcoded, this should be refactored to allow use any name we want.
-          tags.splice(tags.indexOf('YMCA'), 1);
-          tags.splice(tags.indexOf('Camps'), 1);
-          tags.unshift('Camps');
-          tags.unshift('YMCA');
-          tags.forEach(function (tag) {
-            var filter_checked = '';
-            if ($.inArray(tag, this.initial_active_tags) >= 0) {
-              filter_checked = 'selected';
-            }
-            tag_filters_html += '<option value="' + tag + '" ' + filter_checked + '>' + tag + '</option>';
-          }, this);
-          tag_filters_html += '</select>';
-        }
-        else {
-          // Show tags filter as default checkboxes.
-          for (var tag in this.tags) {
-            var filter_checked = '';
-            if ($.inArray(tag, this.initial_active_tags) >= 0) {
-              filter_checked = 'checked="checked"';
-            }
-            var tag_filter_html = '<label class="btn btn-default" for="tag_' + tag + '">';
-            tag_filter_html += '<input autocomplete="off" id="tag_' + tag + '" class="tag_' + tag + '" type="checkbox" value="' + tag + '" ' + filter_checked + '/>' + tag;
-            for (var i = 0; i < this.tags[tag].marker_icons.length; i++) {
-              tag_filter_html += '<img class="tag_icon inline-hidden-sm" src="' + this.tags[tag].marker_icons[i] + '" aria-hidden="true" />';
-            }
-            tag_filter_html += '</label>';
-            tag_filters_html += tag_filter_html;
-          }
-        }
-
-        this.map_controls_el.find('.tag_filters').append(tag_filters_html);
-
-        if (this.tags_style == 'list-checkboxes') {
-          // Init multiselect if used list-checkboxes.
-          $(".tag_filters .multiselect").multiselect({
-            columns: 1,
-            showCheckbox: true,
-            minHeight: 50,
-            texts: {
-              placeholder: 'Select options'
-            },
-            onOptionClick: $.proxy(this.filter_change, this)
-          });
-        }
-        // Add locations autocomplete to search field.
-        var locations = [];
-        this.locations.forEach(function (location) {
-          locations.push(location.name);
-        });
-        this.search_field_el.autocomplete({
-          minLength: 3,
-          source: locations
-        });
-      },
-
-      // Convert string to url format:
-      // remove all non-alphanumeric characters, convert to lowercase,
-      // replace spaces with dashes.
-      encode_to_url_format: function (txt) {
-        return txt
-            .toLowerCase()
-            .replace(/[^\w ]+/g, '')
-            .replace(/ +/g, '-');
-      },
-
-      // Update locations on the map by setting their visibility
-      // and refit the map bounds to the current set of visible locations.
-      draw_map_locations: function () {
-        var locations = this.apply_filters(this.locations);
-
-        // If the location list is empty, don't adjust the map at all.
-        if (locations.length === 0) {
-          if (this.search_center_point !== null) {
-            this.map.setView(this.search_center_point);
-          }
-          return;
-        }
-
-        var bounds = L.latLngBounds([]);
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          bounds.extend(loc.point);
-          loc.marker.addTo(this.map);
-        }
-
-        // Don't zoom in too far on only one marker.
-        if (bounds.getNorthEast().equals(bounds.getSouthWest())) {
-          var extendPoint1 = L.latLng(bounds.getNorthEast().lat + 0.001, bounds.getNorthEast().lng + 0.001);
-          var extendPoint2 = L.latLng(bounds.getNorthEast().lat - 0.001, bounds.getNorthEast().lng - 0.001);
-          bounds.extend(extendPoint1);
-          bounds.extend(extendPoint2);
-        }
-        this.map.fitBounds(bounds, this.fitBoundsOptions);
-      },
-
-      // Updates locations on the map by setting their visibility
-      // to false before drawing.
-      redraw_map_locations: function () {
-        for (var i = 0; i < this.locations.length; i++) {
-          var loc = this.locations[i];
-          loc.marker.removeFrom(this.map);
-        }
-
-        this.draw_map_locations();
-      },
-
-      // Render the list of locations.
-      draw_list_locations: function () {
-        var locations = this.apply_filters(this.locations);
-
-        // Hide all heading locations.
-        for (var l = 0; l < this.locations.length; l++) {
-          if (typeof this.locations[l].element !== 'undefined') {
-            this.locations[l].element.hide();
-            $(this.locations[l].element).parents('.locations-list').find('.location-title').hide();
-          }
-        }
-
-        if (!locations.length) {
-          var message_html = '<div class="col-xs-12 text-center"><p>' +
-              Drupal.t('No locations were found in this area. Please try a different area or increase your search distance.') +
-              '</p></div>';
-          this.messages_el.hide().html(message_html).fadeIn();
-          return;
-        }
-        else {
-          this.messages_el.hide();
-        }
-
-        // Show filtered locations.
-        for (var k = 0; k < locations.length; k++) {
-          if (typeof locations[k].element !== 'undefined') {
-            locations[k].element.show();
-            $(locations[k].element).parents('.locations-list').find('.location-title').show();
-          }
-        }
-      },
-
-      // Generate the HTML for a single location in the list.
-      draw_list_location: function (loc) {
-        return loc.markup;
-      },
-
-      // Generate the HTML for a single location's map detail view
-      draw_map_location: function (loc) {
-        return this.draw_list_location(loc);
-      },
-
-      init_map_locations: function () {
-        var locations = this.locations;
-        var iconOptionsKeys = ['iconSize', 'shadowSize', 'iconAnchor', 'shadowAnchor', 'popupAnchor'];
-
-        for (var i = 0; i < locations.length; i++) {
-          var loc = locations[i];
-          loc.point = L.latLng(loc.lat, loc.lng);
-          var html = '<div class="marker_tooltip">' + this.draw_map_location(loc) + '</div>';
-
-          var icon_options = {
-            iconUrl: loc.icon,
-            iconSize: [32, 42],
-            iconAnchor: [16, 38],
-            popupAnchor: [0, -36]
-            //shadowAnchor: [4, 62],
-            //shadowSize:   [50, 64],
-          };
-
-          $(iconOptionsKeys).each(function (key) {
-            if (typeof loc[key] != 'undefined') {
-              icon_options[key] = loc[key];
-            }
-          });
-
-          var icon = loc.icon ? L.icon(icon_options) : new L.Icon.Default();
-
-          var marker = L.marker(loc.point, {
-            icon: icon
-          });
-          marker.bindPopup(html, { maxWidth: 180 }).openPopup();
-          loc.marker = marker;
-        }
-      },
-
-      draw_search_center: function () {
-        if (this.search_center_point) {
-          this.search_center_marker.setLatLng(this.search_center_point);
-          this.search_center_marker.addTo(this.map);
-        }
-      },
-
-      // Executed every time a checkbox bar filter state changes.
-      bar_filter_change: function (evt) {
-        var evt_target = $(evt.currentTarget);
-        var matching_checkbox = this.map_controls_el.find('input.' + evt_target.attr('class') + '[type=checkbox]');
-
-        // Uncheck all checkboxes.
-        this.map_controls_el.find('input[type=checkbox]').prop('checked', false);
-        this.component_el.find('nav.types input[type=checkbox]').prop('checked', false);
-        this.component_el.find('nav.types label').removeClass('checked');
-
-        // Check the just-clicked one.
-        matching_checkbox.prop('checked', true);
-        evt_target.prop('checked', true);
-
-        evt_target.parents('label').addClass('checked');
-        this.filter_change();
-      }
-    };
-  };
-
-  Drupal.behaviors.openyMap = {
-    attach: function (context, settings) {
-      if (typeof settings.openyMap === 'undefined' || typeof settings.openyMapSettings === 'undefined') {
-        return;
-      }
-
-      var data = settings.openyMap;
-      var map;
-
-      switch (settings.openyMapSettings.engine) {
-        case 'gmaps':
-          map = new Drupal.openyMap();
-          break;
-
-        case 'leaflet':
-        default:
-          map = new Drupal.openyMapLeaflet();
-          map.default_search_location = settings.openyMapSettings.default_location;
-          map.search_icon = settings.openyMapSettings.search_icon;
-          map.search_icon_retina = settings.openyMapSettings.search_icon_retina;
-          switch (settings.openyMapSettings.base_layer) {
-            case 'Esri.WorldStreetMap':
-              map.baseLayer = Drupal.baseLayerEsriWorldStreetMap;
-              break;
-
-            case 'Esri.NatGeoWorldMap':
-              map.baseLayer = Drupal.baseLayerEsriNatGeoWorldMap;
-              break;
-
-            case 'OpenStreetMap.Mapnik':
-              map.baseLayer = Drupal.baseLayerOpenStreetMapMapnik;
-              break;
-
-            case 'Wikimedia':
-              map.baseLayer = Drupal.baseLayerWikimedia;
-              break;
-          }
-          break;
-      }
-
-      $('.locations-list .node--view-mode-teaser').each(function () {
-        var $self = $(this);
-        for (var i = 0; i < data.length; i++) {
-          if (typeof(data[i]) !== 'undefined' && $self.find('.location-item--title')[0].innerText !== 'undefined') {
-            if ($.trim($self.find('.location-item--title')[0].innerText).toLowerCase() == $.trim(data[i].name).toLocaleLowerCase()) {
-              data[i].element = {};
-              data[i].element = $self.parent();
-              data[i].amenities = [];
-              data[i].amenities = ($self.data('amenities'));
-            }
-          }
-        }
-      });
-
-
-      $('.openy-map-canvas', context).once().each(function () {
-        var $canvas = $(this);
-        var timer = setInterval(function () {
-          if (!map.libraryIsLoaded()) {
-            return;
-          }
-
-          map.init({
-            component_el: $canvas.closest('.openy-map-wrapper'),
-            map_data: data,
-            tags_style: $canvas.closest('.location-finder-filters').attr('data-tags-style')
-          });
-
-          // Reset openyMap data (fix for old pins on new map after ajax call).
-          settings.openyMap = [];
-          clearInterval(timer);
-        }, 100);
-      });
-
-    }
-  };
-
-})(jQuery, window, Drupal, drupalSettings);
+!function(t){var e={};function n(r){if(e[r])return e[r].exports;var i=e[r]={i:r,l:!1,exports:{}};return t[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)n.d(r,i,function(e){return t[e]}.bind(null,i));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=158)}([function(t,e,n){var r=n(9),i=n(27),o=n(10),a=n(29),s=n(74),c=function(t,e,n){var l,u,h,f,p=t&c.F,_=t&c.G,d=t&c.S,g=t&c.P,m=t&c.B,v=_?r:d?r[e]||(r[e]={}):(r[e]||{}).prototype,y=_?i:i[e]||(i[e]={}),x=y.prototype||(y.prototype={});for(l in _&&(n=e),n)h=((u=!p&&v&&void 0!==v[l])?v:n)[l],f=m&&u?s(h,r):g&&"function"==typeof h?s(Function.call,h):h,v&&a(v,l,h,t&c.U),y[l]!=h&&o(y,l,f),g&&x[l]!=h&&(x[l]=h)};r.core=i,c.F=1,c.G=2,c.S=4,c.P=8,c.B=16,c.W=32,c.U=64,c.R=128,t.exports=c},function(t,e,n){var r=n(51)("wks"),i=n(50),o=n(9).Symbol,a="function"==typeof o;(t.exports=function(t){return r[t]||(r[t]=a&&o[t]||(a?o:i)("Symbol."+t))}).store=r},function(t,e,n){var r=n(46)("wks"),i=n(26),o=n(5).Symbol,a="function"==typeof o;(t.exports=function(t){return r[t]||(r[t]=a&&o[t]||(a?o:i)("Symbol."+t))}).store=r},function(t,e,n){var r=n(19);t.exports=function(t){if(!r(t))throw TypeError(t+" is not an object!");return t}},function(t,e,n){t.exports=n(131)},function(t,e){var n=t.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(t,e){var n=t.exports={version:"2.6.10"};"number"==typeof __e&&(__e=n)},function(t,e){t.exports=function(t){try{return!!t()}catch(t){return!0}}},function(t,e){var n={}.hasOwnProperty;t.exports=function(t,e){return n.call(t,e)}},function(t,e){var n=t.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(t,e,n){var r=n(28),i=n(72);t.exports=n(15)?function(t,e,n){return r.f(t,e,i(1,n))}:function(t,e,n){return t[e]=n,t}},function(t,e,n){var r=n(12),i=n(24);t.exports=n(13)?function(t,e,n){return r.f(t,e,i(1,n))}:function(t,e,n){return t[e]=n,t}},function(t,e,n){var r=n(16),i=n(62),o=n(39),a=Object.defineProperty;e.f=n(13)?Object.defineProperty:function(t,e,n){if(r(t),e=o(e,!0),r(n),i)try{return a(t,e,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(t[e]=n.value),t}},function(t,e,n){t.exports=!n(23)((function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a}))},function(t,e,n){var r=n(93),i=n(42);t.exports=function(t){return r(i(t))}},function(t,e,n){t.exports=!n(7)((function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a}))},function(t,e,n){var r=n(17);t.exports=function(t){if(!r(t))throw TypeError(t+" is not an object!");return t}},function(t,e){t.exports=function(t){return"object"==typeof t?null!==t:"function"==typeof t}},function(t,e){t.exports={}},function(t,e){t.exports=function(t){return"object"==typeof t?null!==t:"function"==typeof t}},function(t,e){t.exports=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t}},function(t,e,n){var r=n(34),i=Math.min;t.exports=function(t){return t>0?i(r(t),9007199254740991):0}},function(t,e,n){var r=n(20);t.exports=function(t){return Object(r(t))}},function(t,e){t.exports=function(t){try{return!!t()}catch(t){return!0}}},function(t,e){t.exports=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}}},function(t,e){t.exports=!0},function(t,e){var n=0,r=Math.random();t.exports=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++n+r).toString(36))}},function(t,e){var n=t.exports={version:"2.6.10"};"number"==typeof __e&&(__e=n)},function(t,e,n){var r=n(3),i=n(107),o=n(108),a=Object.defineProperty;e.f=n(15)?Object.defineProperty:function(t,e,n){if(r(t),e=o(e,!0),r(n),i)try{return a(t,e,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(t[e]=n.value),t}},function(t,e,n){var r=n(9),i=n(10),o=n(30),a=n(50)("src"),s=n(109),c=(""+s).split("toString");n(27).inspectSource=function(t){return s.call(t)},(t.exports=function(t,e,n,s){var l="function"==typeof n;l&&(o(n,"name")||i(n,"name",e)),t[e]!==n&&(l&&(o(n,a)||i(n,a,t[e]?""+t[e]:c.join(String(e)))),t===r?t[e]=n:s?t[e]?t[e]=n:i(t,e,n):(delete t[e],i(t,e,n)))})(Function.prototype,"toString",(function(){return"function"==typeof this&&this[a]||s.call(this)}))},function(t,e){var n={}.hasOwnProperty;t.exports=function(t,e){return n.call(t,e)}},function(t,e,n){var r=n(111),i=n(77);t.exports=Object.keys||function(t){return r(t,i)}},function(t,e,n){var r=n(75),i=n(20);t.exports=function(t){return r(i(t))}},function(t,e){var n={}.toString;t.exports=function(t){return n.call(t).slice(8,-1)}},function(t,e){var n=Math.ceil,r=Math.floor;t.exports=function(t){return isNaN(t=+t)?0:(t>0?r:n)(t)}},function(t,e,n){"use strict";var r=n(7);t.exports=function(t,e){return!!t&&r((function(){e?t.call(null,(function(){}),1):t.call(null)}))}},function(t,e,n){"use strict";var r=n(121),i=RegExp.prototype.exec;t.exports=function(t,e){var n=t.exec;if("function"==typeof n){var o=n.call(t,e);if("object"!=typeof o)throw new TypeError("RegExp exec method returned something other than an Object or null");return o}if("RegExp"!==r(t))throw new TypeError("RegExp#exec called on incompatible receiver");return i.call(t,e)}},function(t,e,n){"use strict";n(122);var r=n(29),i=n(10),o=n(7),a=n(20),s=n(1),c=n(54),l=s("species"),u=!o((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),h=function(){var t=/(?:)/,e=t.exec;t.exec=function(){return e.apply(this,arguments)};var n="ab".split(t);return 2===n.length&&"a"===n[0]&&"b"===n[1]}();t.exports=function(t,e,n){var f=s(t),p=!o((function(){var e={};return e[f]=function(){return 7},7!=""[t](e)})),_=p?!o((function(){var e=!1,n=/a/;return n.exec=function(){return e=!0,null},"split"===t&&(n.constructor={},n.constructor[l]=function(){return n}),n[f](""),!e})):void 0;if(!p||!_||"replace"===t&&!u||"split"===t&&!h){var d=/./[f],g=n(a,f,""[t],(function(t,e,n,r,i){return e.exec===c?p&&!i?{done:!0,value:d.call(e,n,r)}:{done:!0,value:t.call(n,e,r)}:{done:!1}})),m=g[0],v=g[1];r(String.prototype,t,m),i(RegExp.prototype,f,2==e?function(t,e){return v.call(t,this,e)}:function(t){return v.call(t,this)})}}},function(t,e,n){var r=n(5),i=n(6),o=n(87),a=n(11),s=n(8),c=function(t,e,n){var l,u,h,f=t&c.F,p=t&c.G,_=t&c.S,d=t&c.P,g=t&c.B,m=t&c.W,v=p?i:i[e]||(i[e]={}),y=v.prototype,x=p?r:_?r[e]:(r[e]||{}).prototype;for(l in p&&(n=e),n)(u=!f&&x&&void 0!==x[l])&&s(v,l)||(h=u?x[l]:n[l],v[l]=p&&"function"!=typeof x[l]?n[l]:g&&u?o(h,r):m&&x[l]==h?function(t){var e=function(e,n,r){if(this instanceof t){switch(arguments.length){case 0:return new t;case 1:return new t(e);case 2:return new t(e,n)}return new t(e,n,r)}return t.apply(this,arguments)};return e.prototype=t.prototype,e}(h):d&&"function"==typeof h?o(Function.call,h):h,d&&((v.virtual||(v.virtual={}))[l]=h,t&c.R&&y&&!y[l]&&a(y,l,h)))};c.F=1,c.G=2,c.S=4,c.P=8,c.B=16,c.W=32,c.U=64,c.R=128,t.exports=c},function(t,e,n){var r=n(17);t.exports=function(t,e){if(!r(t))return t;var n,i;if(e&&"function"==typeof(n=t.toString)&&!r(i=n.call(t)))return i;if("function"==typeof(n=t.valueOf)&&!r(i=n.call(t)))return i;if(!e&&"function"==typeof(n=t.toString)&&!r(i=n.call(t)))return i;throw TypeError("Can't convert object to primitive value")}},function(t,e){var n={}.toString;t.exports=function(t){return n.call(t).slice(8,-1)}},function(t,e,n){n(90);for(var r=n(5),i=n(11),o=n(18),a=n(2)("toStringTag"),s="CSSRuleList,CSSStyleDeclaration,CSSValueList,ClientRectList,DOMRectList,DOMStringList,DOMTokenList,DataTransferItemList,FileList,HTMLAllCollection,HTMLCollection,HTMLFormElement,HTMLSelectElement,MediaList,MimeTypeArray,NamedNodeMap,NodeList,PaintRequestList,Plugin,PluginArray,SVGLengthList,SVGNumberList,SVGPathSegList,SVGPointList,SVGStringList,SVGTransformList,SourceBufferList,StyleSheetList,TextTrackCueList,TextTrackList,TouchList".split(","),c=0;c<s.length;c++){var l=s[c],u=r[l],h=u&&u.prototype;h&&!h[a]&&i(h,a,l),o[l]=o.Array}},function(t,e){t.exports=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t}},function(t,e,n){var r=n(68),i=n(47);t.exports=Object.keys||function(t){return r(t,i)}},function(t,e){var n=Math.ceil,r=Math.floor;t.exports=function(t){return isNaN(t=+t)?0:(t>0?r:n)(t)}},function(t,e,n){var r=n(46)("keys"),i=n(26);t.exports=function(t){return r[t]||(r[t]=i(t))}},function(t,e,n){var r=n(6),i=n(5),o=i["__core-js_shared__"]||(i["__core-js_shared__"]={});(t.exports=function(t,e){return o[t]||(o[t]=void 0!==e?e:{})})("versions",[]).push({version:r.version,mode:n(25)?"pure":"global",copyright:"© 2019 Denis Pushkarev (zloirock.ru)"})},function(t,e){t.exports="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(",")},function(t,e,n){var r=n(12).f,i=n(8),o=n(2)("toStringTag");t.exports=function(t,e,n){t&&!i(t=n?t:t.prototype,o)&&r(t,o,{configurable:!0,value:e})}},function(t,e,n){"use strict";var r=n(101)(!0);n(65)(String,"String",(function(t){this._t=String(t),this._i=0}),(function(){var t,e=this._t,n=this._i;return n>=e.length?{value:void 0,done:!0}:(t=r(e,n),this._i+=t.length,{value:t,done:!1})}))},function(t,e){var n=0,r=Math.random();t.exports=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++n+r).toString(36))}},function(t,e,n){var r=n(27),i=n(9),o=i["__core-js_shared__"]||(i["__core-js_shared__"]={});(t.exports=function(t,e){return o[t]||(o[t]=void 0!==e?e:{})})("versions",[]).push({version:r.version,mode:n(73)?"pure":"global",copyright:"© 2019 Denis Pushkarev (zloirock.ru)"})},function(t,e){t.exports=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t}},function(t,e,n){var r=n(51)("keys"),i=n(50);t.exports=function(t){return r[t]||(r[t]=i(t))}},function(t,e,n){"use strict";var r,i,o=n(123),a=RegExp.prototype.exec,s=String.prototype.replace,c=a,l=(r=/a/,i=/b*/g,a.call(r,"a"),a.call(i,"a"),0!==r.lastIndex||0!==i.lastIndex),u=void 0!==/()??/.exec("")[1];(l||u)&&(c=function(t){var e,n,r,i,c=this;return u&&(n=new RegExp("^"+c.source+"$(?!\\s)",o.call(c))),l&&(e=c.lastIndex),r=a.call(c,t),l&&r&&(c.lastIndex=c.global?r.index+r[0].length:e),u&&r&&r.length>1&&s.call(r[0],n,(function(){for(i=1;i<arguments.length-2;i++)void 0===arguments[i]&&(r[i]=void 0)})),r}),t.exports=c},function(t,e,n){"use strict";var r=n(127)(!0);t.exports=function(t,e,n){return e+(n?r(t,e).length:1)}},function(t,e,n){e.f=n(2)},function(t,e,n){var r=n(5),i=n(6),o=n(25),a=n(56),s=n(12).f;t.exports=function(t){var e=i.Symbol||(i.Symbol=o?{}:r.Symbol||{});"_"==t.charAt(0)||t in e||s(e,t,{value:a.f(t)})}},function(t,e){e.f={}.propertyIsEnumerable},function(t,e){t.exports={}},function(t,e,n){var r=n(74),i=n(75),o=n(22),a=n(21),s=n(151);t.exports=function(t,e){var n=1==t,c=2==t,l=3==t,u=4==t,h=6==t,f=5==t||h,p=e||s;return function(e,s,_){for(var d,g,m=o(e),v=i(m),y=r(s,_,3),x=a(v.length),b=0,w=n?p(e,x):c?p(e,0):void 0;x>b;b++)if((f||b in v)&&(g=y(d=v[b],b,m),t))if(n)w[b]=g;else if(g)switch(t){case 3:return!0;case 5:return d;case 6:return b;case 2:w.push(d)}else if(u)return!1;return h?-1:l||u?u:w}}},function(t,e,n){t.exports=n(130)},function(t,e,n){t.exports=!n(13)&&!n(23)((function(){return 7!=Object.defineProperty(n(63)("div"),"a",{get:function(){return 7}}).a}))},function(t,e,n){var r=n(17),i=n(5).document,o=r(i)&&r(i.createElement);t.exports=function(t){return o?i.createElement(t):{}}},function(t,e,n){var r=n(40);t.exports=Array.isArray||function(t){return"Array"==r(t)}},function(t,e,n){"use strict";var r=n(25),i=n(38),o=n(66),a=n(11),s=n(18),c=n(94),l=n(48),u=n(100),h=n(2)("iterator"),f=!([].keys&&"next"in[].keys()),p=function(){return this};t.exports=function(t,e,n,_,d,g,m){c(n,e,_);var v,y,x,b=function(t){if(!f&&t in O)return O[t];switch(t){case"keys":case"values":return function(){return new n(this,t)}}return function(){return new n(this,t)}},w=e+" Iterator",k="values"==d,S=!1,O=t.prototype,L=O[h]||O["@@iterator"]||d&&O[d],M=L||b(d),E=d?k?b("entries"):M:void 0,P="Array"==e&&O.entries||L;if(P&&(x=u(P.call(new t)))!==Object.prototype&&x.next&&(l(x,w,!0),r||"function"==typeof x[h]||a(x,h,p)),k&&L&&"values"!==L.name&&(S=!0,M=function(){return L.call(this)}),r&&!m||!f&&!S&&O[h]||a(O,h,M),s[e]=M,s[w]=p,d)if(v={values:k?M:b("values"),keys:g?M:b("keys"),entries:E},m)for(y in v)y in O||o(O,y,v[y]);else i(i.P+i.F*(f||S),e,v);return v}},function(t,e,n){t.exports=n(11)},function(t,e,n){var r=n(16),i=n(95),o=n(47),a=n(45)("IE_PROTO"),s=function(){},c=function(){var t,e=n(63)("iframe"),r=o.length;for(e.style.display="none",n(99).appendChild(e),e.src="javascript:",(t=e.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),c=t.F;r--;)delete c.prototype[o[r]];return c()};t.exports=Object.create||function(t,e){var n;return null!==t?(s.prototype=r(t),n=new s,s.prototype=null,n[a]=t):n=c(),void 0===e?n:i(n,e)}},function(t,e,n){var r=n(8),i=n(14),o=n(96)(!1),a=n(45)("IE_PROTO");t.exports=function(t,e){var n,s=i(t),c=0,l=[];for(n in s)n!=a&&r(s,n)&&l.push(n);for(;e.length>c;)r(s,n=e[c++])&&(~o(l,n)||l.push(n));return l}},function(t,e,n){var r=n(42);t.exports=function(t){return Object(r(t))}},function(t,e,n){var r=n(40),i=n(2)("toStringTag"),o="Arguments"==r(function(){return arguments}());t.exports=function(t){var e,n,a;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),i))?n:o?r(e):"Object"==(a=r(e))&&"function"==typeof e.callee?"Arguments":a}},function(t,e,n){var r=n(19),i=n(9).document,o=r(i)&&r(i.createElement);t.exports=function(t){return o?i.createElement(t):{}}},function(t,e){t.exports=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}}},function(t,e){t.exports=!1},function(t,e,n){var r=n(52);t.exports=function(t,e,n){if(r(t),void 0===e)return t;switch(n){case 1:return function(n){return t.call(e,n)};case 2:return function(n,r){return t.call(e,n,r)};case 3:return function(n,r,i){return t.call(e,n,r,i)}}return function(){return t.apply(e,arguments)}}},function(t,e,n){var r=n(33);t.exports=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==r(t)?t.split(""):Object(t)}},function(t,e,n){var r=n(32),i=n(21),o=n(112);t.exports=function(t){return function(e,n,a){var s,c=r(e),l=i(c.length),u=o(a,l);if(t&&n!=n){for(;l>u;)if((s=c[u++])!=s)return!0}else for(;l>u;u++)if((t||u in c)&&c[u]===n)return t||u||0;return!t&&-1}}},function(t,e){t.exports="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(",")},function(t,e){e.f=Object.getOwnPropertySymbols},function(t,e,n){var r=n(68),i=n(47).concat("length","prototype");e.f=Object.getOwnPropertyNames||function(t){return r(t,i)}},function(t,e,n){var r=n(1)("unscopables"),i=Array.prototype;null==i[r]&&n(10)(i,r,{}),t.exports=function(t){i[r][t]=!0}},function(t,e,n){var r=n(28).f,i=n(30),o=n(1)("toStringTag");t.exports=function(t,e,n){t&&!i(t=n?t:t.prototype,o)&&r(t,o,{configurable:!0,value:e})}},function(t,e,n){t.exports=n(85)},function(t,e,n){t.exports=n(89)},function(t,e,n){t.exports=n(104)},function(t,e,n){n(86),t.exports=n(6).Array.isArray},function(t,e,n){var r=n(38);r(r.S,"Array",{isArray:n(64)})},function(t,e,n){var r=n(88);t.exports=function(t,e,n){if(r(t),void 0===e)return t;switch(n){case 1:return function(n){return t.call(e,n)};case 2:return function(n,r){return t.call(e,n,r)};case 3:return function(n,r,i){return t.call(e,n,r,i)}}return function(){return t.apply(e,arguments)}}},function(t,e){t.exports=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t}},function(t,e,n){n(41),n(49),t.exports=n(102)},function(t,e,n){"use strict";var r=n(91),i=n(92),o=n(18),a=n(14);t.exports=n(65)(Array,"Array",(function(t,e){this._t=a(t),this._i=0,this._k=e}),(function(){var t=this._t,e=this._k,n=this._i++;return!t||n>=t.length?(this._t=void 0,i(1)):i(0,"keys"==e?n:"values"==e?t[n]:[n,t[n]])}),"values"),o.Arguments=o.Array,r("keys"),r("values"),r("entries")},function(t,e){t.exports=function(){}},function(t,e){t.exports=function(t,e){return{value:e,done:!!t}}},function(t,e,n){var r=n(40);t.exports=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==r(t)?t.split(""):Object(t)}},function(t,e,n){"use strict";var r=n(67),i=n(24),o=n(48),a={};n(11)(a,n(2)("iterator"),(function(){return this})),t.exports=function(t,e,n){t.prototype=r(a,{next:i(1,n)}),o(t,e+" Iterator")}},function(t,e,n){var r=n(12),i=n(16),o=n(43);t.exports=n(13)?Object.defineProperties:function(t,e){i(t);for(var n,a=o(e),s=a.length,c=0;s>c;)r.f(t,n=a[c++],e[n]);return t}},function(t,e,n){var r=n(14),i=n(97),o=n(98);t.exports=function(t){return function(e,n,a){var s,c=r(e),l=i(c.length),u=o(a,l);if(t&&n!=n){for(;l>u;)if((s=c[u++])!=s)return!0}else for(;l>u;u++)if((t||u in c)&&c[u]===n)return t||u||0;return!t&&-1}}},function(t,e,n){var r=n(44),i=Math.min;t.exports=function(t){return t>0?i(r(t),9007199254740991):0}},function(t,e,n){var r=n(44),i=Math.max,o=Math.min;t.exports=function(t,e){return(t=r(t))<0?i(t+e,0):o(t,e)}},function(t,e,n){var r=n(5).document;t.exports=r&&r.documentElement},function(t,e,n){var r=n(8),i=n(69),o=n(45)("IE_PROTO"),a=Object.prototype;t.exports=Object.getPrototypeOf||function(t){return t=i(t),r(t,o)?t[o]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?a:null}},function(t,e,n){var r=n(44),i=n(42);t.exports=function(t){return function(e,n){var o,a,s=String(i(e)),c=r(n),l=s.length;return c<0||c>=l?t?"":void 0:(o=s.charCodeAt(c))<55296||o>56319||c+1===l||(a=s.charCodeAt(c+1))<56320||a>57343?t?s.charAt(c):o:t?s.slice(c,c+2):a-56320+(o-55296<<10)+65536}}},function(t,e,n){var r=n(16),i=n(103);t.exports=n(6).getIterator=function(t){var e=i(t);if("function"!=typeof e)throw TypeError(t+" is not iterable!");return r(e.call(t))}},function(t,e,n){var r=n(70),i=n(2)("iterator"),o=n(18);t.exports=n(6).getIteratorMethod=function(t){if(null!=t)return t[i]||t["@@iterator"]||o[r(t)]}},function(t,e,n){n(41),n(49),t.exports=n(105)},function(t,e,n){var r=n(70),i=n(2)("iterator"),o=n(18);t.exports=n(6).isIterable=function(t){var e=Object(t);return void 0!==e[i]||"@@iterator"in e||o.hasOwnProperty(r(e))}},function(t,e,n){var r=n(0),i=n(110)(!0);r(r.S,"Object",{entries:function(t){return i(t)}})},function(t,e,n){t.exports=!n(15)&&!n(7)((function(){return 7!=Object.defineProperty(n(71)("div"),"a",{get:function(){return 7}}).a}))},function(t,e,n){var r=n(19);t.exports=function(t,e){if(!r(t))return t;var n,i;if(e&&"function"==typeof(n=t.toString)&&!r(i=n.call(t)))return i;if("function"==typeof(n=t.valueOf)&&!r(i=n.call(t)))return i;if(!e&&"function"==typeof(n=t.toString)&&!r(i=n.call(t)))return i;throw TypeError("Can't convert object to primitive value")}},function(t,e,n){t.exports=n(51)("native-function-to-string",Function.toString)},function(t,e,n){var r=n(15),i=n(31),o=n(32),a=n(113).f;t.exports=function(t){return function(e){for(var n,s=o(e),c=i(s),l=c.length,u=0,h=[];l>u;)n=c[u++],r&&!a.call(s,n)||h.push(t?[n,s[n]]:s[n]);return h}}},function(t,e,n){var r=n(30),i=n(32),o=n(76)(!1),a=n(53)("IE_PROTO");t.exports=function(t,e){var n,s=i(t),c=0,l=[];for(n in s)n!=a&&r(s,n)&&l.push(n);for(;e.length>c;)r(s,n=e[c++])&&(~o(l,n)||l.push(n));return l}},function(t,e,n){var r=n(34),i=Math.max,o=Math.min;t.exports=function(t,e){return(t=r(t))<0?i(t+e,0):o(t,e)}},function(t,e){e.f={}.propertyIsEnumerable},function(t,e,n){"use strict";n(115)("trim",(function(t){return function(){return t(this,3)}}))},function(t,e,n){var r=n(0),i=n(20),o=n(7),a=n(116),s="["+a+"]",c=RegExp("^"+s+s+"*"),l=RegExp(s+s+"*$"),u=function(t,e,n){var i={},s=o((function(){return!!a[t]()||"​"!="​"[t]()})),c=i[t]=s?e(h):a[t];n&&(i[n]=c),r(r.P+r.F*s,"String",i)},h=u.trim=function(t,e){return t=String(i(t)),1&e&&(t=t.replace(c,"")),2&e&&(t=t.replace(l,"")),t};t.exports=u},function(t,e){t.exports="\t\n\v\f\r   ᠎             　\u2028\u2029\ufeff"},function(t,e,n){"use strict";var r=n(0),i=n(76)(!1),o=[].indexOf,a=!!o&&1/[1].indexOf(1,-0)<0;r(r.P+r.F*(a||!n(35)(o)),"Array",{indexOf:function(t){return a?o.apply(this,arguments)||0:i(this,t,arguments[1])}})},function(t,e,n){"use strict";var r=n(0),i=n(52),o=n(22),a=n(7),s=[].sort,c=[1,2,3];r(r.P+r.F*(a((function(){c.sort(void 0)}))||!a((function(){c.sort(null)}))||!n(35)(s)),"Array",{sort:function(t){return void 0===t?s.call(o(this)):s.call(o(this),i(t))}})},function(t,e,n){"use strict";var r=n(3),i=n(120),o=n(36);n(37)("search",1,(function(t,e,n,a){return[function(n){var r=t(this),i=null==n?void 0:n[e];return void 0!==i?i.call(n,r):new RegExp(n)[e](String(r))},function(t){var e=a(n,t,this);if(e.done)return e.value;var s=r(t),c=String(this),l=s.lastIndex;i(l,0)||(s.lastIndex=0);var u=o(s,c);return i(s.lastIndex,l)||(s.lastIndex=l),null===u?-1:u.index}]}))},function(t,e){t.exports=Object.is||function(t,e){return t===e?0!==t||1/t==1/e:t!=t&&e!=e}},function(t,e,n){var r=n(33),i=n(1)("toStringTag"),o="Arguments"==r(function(){return arguments}());t.exports=function(t){var e,n,a;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),i))?n:o?r(e):"Object"==(a=r(e))&&"function"==typeof e.callee?"Arguments":a}},function(t,e,n){"use strict";var r=n(54);n(0)({target:"RegExp",proto:!0,forced:r!==/./.exec},{exec:r})},function(t,e,n){"use strict";var r=n(3);t.exports=function(){var t=r(this),e="";return t.global&&(e+="g"),t.ignoreCase&&(e+="i"),t.multiline&&(e+="m"),t.unicode&&(e+="u"),t.sticky&&(e+="y"),e}},function(t,e,n){"use strict";var r=n(125),i=n(3),o=n(126),a=n(55),s=n(21),c=n(36),l=n(54),u=n(7),h=Math.min,f=[].push,p=!u((function(){RegExp(4294967295,"y")}));n(37)("split",2,(function(t,e,n,u){var _;return _="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(t,e){var i=String(this);if(void 0===t&&0===e)return[];if(!r(t))return n.call(i,t,e);for(var o,a,s,c=[],u=(t.ignoreCase?"i":"")+(t.multiline?"m":"")+(t.unicode?"u":"")+(t.sticky?"y":""),h=0,p=void 0===e?4294967295:e>>>0,_=new RegExp(t.source,u+"g");(o=l.call(_,i))&&!((a=_.lastIndex)>h&&(c.push(i.slice(h,o.index)),o.length>1&&o.index<i.length&&f.apply(c,o.slice(1)),s=o[0].length,h=a,c.length>=p));)_.lastIndex===o.index&&_.lastIndex++;return h===i.length?!s&&_.test("")||c.push(""):c.push(i.slice(h)),c.length>p?c.slice(0,p):c}:"0".split(void 0,0).length?function(t,e){return void 0===t&&0===e?[]:n.call(this,t,e)}:n,[function(n,r){var i=t(this),o=null==n?void 0:n[e];return void 0!==o?o.call(n,i,r):_.call(String(i),n,r)},function(t,e){var r=u(_,t,this,e,_!==n);if(r.done)return r.value;var l=i(t),f=String(this),d=o(l,RegExp),g=l.unicode,m=(l.ignoreCase?"i":"")+(l.multiline?"m":"")+(l.unicode?"u":"")+(p?"y":"g"),v=new d(p?l:"^(?:"+l.source+")",m),y=void 0===e?4294967295:e>>>0;if(0===y)return[];if(0===f.length)return null===c(v,f)?[f]:[];for(var x=0,b=0,w=[];b<f.length;){v.lastIndex=p?b:0;var k,S=c(v,p?f:f.slice(b));if(null===S||(k=h(s(v.lastIndex+(p?0:b)),f.length))===x)b=a(f,b,g);else{if(w.push(f.slice(x,b)),w.length===y)return w;for(var O=1;O<=S.length-1;O++)if(w.push(S[O]),w.length===y)return w;b=x=k}}return w.push(f.slice(x)),w}]}))},function(t,e,n){var r=n(19),i=n(33),o=n(1)("match");t.exports=function(t){var e;return r(t)&&(void 0!==(e=t[o])?!!e:"RegExp"==i(t))}},function(t,e,n){var r=n(3),i=n(52),o=n(1)("species");t.exports=function(t,e){var n,a=r(t).constructor;return void 0===a||null==(n=r(a)[o])?e:i(n)}},function(t,e,n){var r=n(34),i=n(20);t.exports=function(t){return function(e,n){var o,a,s=String(i(e)),c=r(n),l=s.length;return c<0||c>=l?t?"":void 0:(o=s.charCodeAt(c))<55296||o>56319||c+1===l||(a=s.charCodeAt(c+1))<56320||a>57343?t?s.charAt(c):o:t?s.slice(c,c+2):a-56320+(o-55296<<10)+65536}}},function(t,e,n){var r=n(22),i=n(31);n(129)("keys",(function(){return function(t){return i(r(t))}}))},function(t,e,n){var r=n(0),i=n(27),o=n(7);t.exports=function(t,e){var n=(i.Object||{})[t]||Object[t],a={};a[t]=e(n),r(r.S+r.F*o((function(){n(1)})),"Object",a)}},function(t,e,n){n(49),n(41),t.exports=n(56).f("iterator")},function(t,e,n){n(132),n(137),n(138),n(139),t.exports=n(6).Symbol},function(t,e,n){"use strict";var r=n(5),i=n(8),o=n(13),a=n(38),s=n(66),c=n(133).KEY,l=n(23),u=n(46),h=n(48),f=n(26),p=n(2),_=n(56),d=n(57),g=n(134),m=n(64),v=n(16),y=n(17),x=n(69),b=n(14),w=n(39),k=n(24),S=n(67),O=n(135),L=n(136),M=n(78),E=n(12),P=n(43),j=L.f,A=E.f,T=O.f,C=r.Symbol,I=r.JSON,N=I&&I.stringify,R=p("_hidden"),F=p("toPrimitive"),W={}.propertyIsEnumerable,G=u("symbol-registry"),V=u("symbols"),B=u("op-symbols"),z=Object.prototype,D="function"==typeof C&&!!M.f,U=r.QObject,q=!U||!U.prototype||!U.prototype.findChild,H=o&&l((function(){return 7!=S(A({},"a",{get:function(){return A(this,"a",{value:7}).a}})).a}))?function(t,e,n){var r=j(z,e);r&&delete z[e],A(t,e,n),r&&t!==z&&A(z,e,r)}:A,$=function(t){var e=V[t]=S(C.prototype);return e._k=t,e},J=D&&"symbol"==typeof C.iterator?function(t){return"symbol"==typeof t}:function(t){return t instanceof C},Y=function(t,e,n){return t===z&&Y(B,e,n),v(t),e=w(e,!0),v(n),i(V,e)?(n.enumerable?(i(t,R)&&t[R][e]&&(t[R][e]=!1),n=S(n,{enumerable:k(0,!1)})):(i(t,R)||A(t,R,k(1,{})),t[R][e]=!0),H(t,e,n)):A(t,e,n)},Z=function(t,e){v(t);for(var n,r=g(e=b(e)),i=0,o=r.length;o>i;)Y(t,n=r[i++],e[n]);return t},K=function(t){var e=W.call(this,t=w(t,!0));return!(this===z&&i(V,t)&&!i(B,t))&&(!(e||!i(this,t)||!i(V,t)||i(this,R)&&this[R][t])||e)},Q=function(t,e){if(t=b(t),e=w(e,!0),t!==z||!i(V,e)||i(B,e)){var n=j(t,e);return!n||!i(V,e)||i(t,R)&&t[R][e]||(n.enumerable=!0),n}},X=function(t){for(var e,n=T(b(t)),r=[],o=0;n.length>o;)i(V,e=n[o++])||e==R||e==c||r.push(e);return r},tt=function(t){for(var e,n=t===z,r=T(n?B:b(t)),o=[],a=0;r.length>a;)!i(V,e=r[a++])||n&&!i(z,e)||o.push(V[e]);return o};D||(s((C=function(){if(this instanceof C)throw TypeError("Symbol is not a constructor!");var t=f(arguments.length>0?arguments[0]:void 0),e=function(n){this===z&&e.call(B,n),i(this,R)&&i(this[R],t)&&(this[R][t]=!1),H(this,t,k(1,n))};return o&&q&&H(z,t,{configurable:!0,set:e}),$(t)}).prototype,"toString",(function(){return this._k})),L.f=Q,E.f=Y,n(79).f=O.f=X,n(58).f=K,M.f=tt,o&&!n(25)&&s(z,"propertyIsEnumerable",K,!0),_.f=function(t){return $(p(t))}),a(a.G+a.W+a.F*!D,{Symbol:C});for(var et="hasInstance,isConcatSpreadable,iterator,match,replace,search,species,split,toPrimitive,toStringTag,unscopables".split(","),nt=0;et.length>nt;)p(et[nt++]);for(var rt=P(p.store),it=0;rt.length>it;)d(rt[it++]);a(a.S+a.F*!D,"Symbol",{for:function(t){return i(G,t+="")?G[t]:G[t]=C(t)},keyFor:function(t){if(!J(t))throw TypeError(t+" is not a symbol!");for(var e in G)if(G[e]===t)return e},useSetter:function(){q=!0},useSimple:function(){q=!1}}),a(a.S+a.F*!D,"Object",{create:function(t,e){return void 0===e?S(t):Z(S(t),e)},defineProperty:Y,defineProperties:Z,getOwnPropertyDescriptor:Q,getOwnPropertyNames:X,getOwnPropertySymbols:tt});var ot=l((function(){M.f(1)}));a(a.S+a.F*ot,"Object",{getOwnPropertySymbols:function(t){return M.f(x(t))}}),I&&a(a.S+a.F*(!D||l((function(){var t=C();return"[null]"!=N([t])||"{}"!=N({a:t})||"{}"!=N(Object(t))}))),"JSON",{stringify:function(t){for(var e,n,r=[t],i=1;arguments.length>i;)r.push(arguments[i++]);if(n=e=r[1],(y(e)||void 0!==t)&&!J(t))return m(e)||(e=function(t,e){if("function"==typeof n&&(e=n.call(this,t,e)),!J(e))return e}),r[1]=e,N.apply(I,r)}}),C.prototype[F]||n(11)(C.prototype,F,C.prototype.valueOf),h(C,"Symbol"),h(Math,"Math",!0),h(r.JSON,"JSON",!0)},function(t,e,n){var r=n(26)("meta"),i=n(17),o=n(8),a=n(12).f,s=0,c=Object.isExtensible||function(){return!0},l=!n(23)((function(){return c(Object.preventExtensions({}))})),u=function(t){a(t,r,{value:{i:"O"+ ++s,w:{}}})},h=t.exports={KEY:r,NEED:!1,fastKey:function(t,e){if(!i(t))return"symbol"==typeof t?t:("string"==typeof t?"S":"P")+t;if(!o(t,r)){if(!c(t))return"F";if(!e)return"E";u(t)}return t[r].i},getWeak:function(t,e){if(!o(t,r)){if(!c(t))return!0;if(!e)return!1;u(t)}return t[r].w},onFreeze:function(t){return l&&h.NEED&&c(t)&&!o(t,r)&&u(t),t}}},function(t,e,n){var r=n(43),i=n(78),o=n(58);t.exports=function(t){var e=r(t),n=i.f;if(n)for(var a,s=n(t),c=o.f,l=0;s.length>l;)c.call(t,a=s[l++])&&e.push(a);return e}},function(t,e,n){var r=n(14),i=n(79).f,o={}.toString,a="object"==typeof window&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[];t.exports.f=function(t){return a&&"[object Window]"==o.call(t)?function(t){try{return i(t)}catch(t){return a.slice()}}(t):i(r(t))}},function(t,e,n){var r=n(58),i=n(24),o=n(14),a=n(39),s=n(8),c=n(62),l=Object.getOwnPropertyDescriptor;e.f=n(13)?l:function(t,e){if(t=o(t),e=a(e,!0),c)try{return l(t,e)}catch(t){}if(s(t,e))return i(!r.f.call(t,e),t[e])}},function(t,e){},function(t,e,n){n(57)("asyncIterator")},function(t,e,n){n(57)("observable")},function(t,e,n){var r=n(28).f,i=Function.prototype,o=/^\s*function ([^ (]*)/;"name"in i||n(15)&&r(i,"name",{configurable:!0,get:function(){try{return(""+this).match(o)[1]}catch(t){return""}}})},function(t,e,n){for(var r=n(142),i=n(31),o=n(29),a=n(9),s=n(10),c=n(59),l=n(1),u=l("iterator"),h=l("toStringTag"),f=c.Array,p={CSSRuleList:!0,CSSStyleDeclaration:!1,CSSValueList:!1,ClientRectList:!1,DOMRectList:!1,DOMStringList:!1,DOMTokenList:!0,DataTransferItemList:!1,FileList:!1,HTMLAllCollection:!1,HTMLCollection:!1,HTMLFormElement:!1,HTMLSelectElement:!1,MediaList:!0,MimeTypeArray:!1,NamedNodeMap:!1,NodeList:!0,PaintRequestList:!1,Plugin:!1,PluginArray:!1,SVGLengthList:!1,SVGNumberList:!1,SVGPathSegList:!1,SVGPointList:!1,SVGStringList:!1,SVGTransformList:!1,SourceBufferList:!1,StyleSheetList:!0,TextTrackCueList:!1,TextTrackList:!1,TouchList:!1},_=i(p),d=0;d<_.length;d++){var g,m=_[d],v=p[m],y=a[m],x=y&&y.prototype;if(x&&(x[u]||s(x,u,f),x[h]||s(x,h,m),c[m]=f,v))for(g in r)x[g]||o(x,g,r[g],!0)}},function(t,e,n){"use strict";var r=n(80),i=n(143),o=n(59),a=n(32);t.exports=n(144)(Array,"Array",(function(t,e){this._t=a(t),this._i=0,this._k=e}),(function(){var t=this._t,e=this._k,n=this._i++;return!t||n>=t.length?(this._t=void 0,i(1)):i(0,"keys"==e?n:"values"==e?t[n]:[n,t[n]])}),"values"),o.Arguments=o.Array,r("keys"),r("values"),r("entries")},function(t,e){t.exports=function(t,e){return{value:e,done:!!t}}},function(t,e,n){"use strict";var r=n(73),i=n(0),o=n(29),a=n(10),s=n(59),c=n(145),l=n(81),u=n(149),h=n(1)("iterator"),f=!([].keys&&"next"in[].keys()),p=function(){return this};t.exports=function(t,e,n,_,d,g,m){c(n,e,_);var v,y,x,b=function(t){if(!f&&t in O)return O[t];switch(t){case"keys":case"values":return function(){return new n(this,t)}}return function(){return new n(this,t)}},w=e+" Iterator",k="values"==d,S=!1,O=t.prototype,L=O[h]||O["@@iterator"]||d&&O[d],M=L||b(d),E=d?k?b("entries"):M:void 0,P="Array"==e&&O.entries||L;if(P&&(x=u(P.call(new t)))!==Object.prototype&&x.next&&(l(x,w,!0),r||"function"==typeof x[h]||a(x,h,p)),k&&L&&"values"!==L.name&&(S=!0,M=function(){return L.call(this)}),r&&!m||!f&&!S&&O[h]||a(O,h,M),s[e]=M,s[w]=p,d)if(v={values:k?M:b("values"),keys:g?M:b("keys"),entries:E},m)for(y in v)y in O||o(O,y,v[y]);else i(i.P+i.F*(f||S),e,v);return v}},function(t,e,n){"use strict";var r=n(146),i=n(72),o=n(81),a={};n(10)(a,n(1)("iterator"),(function(){return this})),t.exports=function(t,e,n){t.prototype=r(a,{next:i(1,n)}),o(t,e+" Iterator")}},function(t,e,n){var r=n(3),i=n(147),o=n(77),a=n(53)("IE_PROTO"),s=function(){},c=function(){var t,e=n(71)("iframe"),r=o.length;for(e.style.display="none",n(148).appendChild(e),e.src="javascript:",(t=e.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),c=t.F;r--;)delete c.prototype[o[r]];return c()};t.exports=Object.create||function(t,e){var n;return null!==t?(s.prototype=r(t),n=new s,s.prototype=null,n[a]=t):n=c(),void 0===e?n:i(n,e)}},function(t,e,n){var r=n(28),i=n(3),o=n(31);t.exports=n(15)?Object.defineProperties:function(t,e){i(t);for(var n,a=o(e),s=a.length,c=0;s>c;)r.f(t,n=a[c++],e[n]);return t}},function(t,e,n){var r=n(9).document;t.exports=r&&r.documentElement},function(t,e,n){var r=n(30),i=n(22),o=n(53)("IE_PROTO"),a=Object.prototype;t.exports=Object.getPrototypeOf||function(t){return t=i(t),r(t,o)?t[o]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?a:null}},function(t,e,n){"use strict";var r=n(0),i=n(60)(0),o=n(35)([].forEach,!0);r(r.P+r.F*!o,"Array",{forEach:function(t){return i(this,t,arguments[1])}})},function(t,e,n){var r=n(152);t.exports=function(t,e){return new(r(t))(e)}},function(t,e,n){var r=n(19),i=n(153),o=n(1)("species");t.exports=function(t){var e;return i(t)&&("function"!=typeof(e=t.constructor)||e!==Array&&!i(e.prototype)||(e=void 0),r(e)&&null===(e=e[o])&&(e=void 0)),void 0===e?Array:e}},function(t,e,n){var r=n(33);t.exports=Array.isArray||function(t){return"Array"==r(t)}},function(t,e,n){"use strict";var r=n(0),i=n(60)(1);r(r.P+r.F*!n(35)([].map,!0),"Array",{map:function(t){return i(this,t,arguments[1])}})},function(t,e,n){"use strict";var r=n(3),i=n(22),o=n(21),a=n(34),s=n(55),c=n(36),l=Math.max,u=Math.min,h=Math.floor,f=/\$([$&`']|\d\d?|<[^>]*>)/g,p=/\$([$&`']|\d\d?)/g;n(37)("replace",2,(function(t,e,n,_){return[function(r,i){var o=t(this),a=null==r?void 0:r[e];return void 0!==a?a.call(r,o,i):n.call(String(o),r,i)},function(t,e){var i=_(n,t,this,e);if(i.done)return i.value;var h=r(t),f=String(this),p="function"==typeof e;p||(e=String(e));var g=h.global;if(g){var m=h.unicode;h.lastIndex=0}for(var v=[];;){var y=c(h,f);if(null===y)break;if(v.push(y),!g)break;""===String(y[0])&&(h.lastIndex=s(f,o(h.lastIndex),m))}for(var x,b="",w=0,k=0;k<v.length;k++){y=v[k];for(var S=String(y[0]),O=l(u(a(y.index),f.length),0),L=[],M=1;M<y.length;M++)L.push(void 0===(x=y[M])?x:String(x));var E=y.groups;if(p){var P=[S].concat(L,O,f);void 0!==E&&P.push(E);var j=String(e.apply(void 0,P))}else j=d(S,f,O,L,E,e);O>=w&&(b+=f.slice(w,O)+j,w=O+S.length)}return b+f.slice(w)}];function d(t,e,r,o,a,s){var c=r+t.length,l=o.length,u=p;return void 0!==a&&(a=i(a),u=f),n.call(s,u,(function(n,i){var s;switch(i.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,r);case"'":return e.slice(c);case"<":s=a[i.slice(1,-1)];break;default:var u=+i;if(0===u)return n;if(u>l){var f=h(u/10);return 0===f?n:f<=l?void 0===o[f-1]?i.charAt(1):o[f-1]+i.charAt(1):n}s=o[u-1]}return void 0===s?"":s}))}}))},function(t,e,n){"use strict";var r=n(3),i=n(21),o=n(55),a=n(36);n(37)("match",1,(function(t,e,n,s){return[function(n){var r=t(this),i=null==n?void 0:n[e];return void 0!==i?i.call(n,r):new RegExp(n)[e](String(r))},function(t){var e=s(n,t,this);if(e.done)return e.value;var c=r(t),l=String(this);if(!c.global)return a(c,l);var u=c.unicode;c.lastIndex=0;for(var h,f=[],p=0;null!==(h=a(c,l));){var _=String(h[0]);f[p]=_,""===_&&(c.lastIndex=o(l,i(c.lastIndex),u)),p++}return 0===p?null:f}]}))},function(t,e,n){"use strict";var r=n(0),i=n(60)(5),o=!0;"find"in[]&&Array(1).find((function(){o=!1})),r(r.P+r.F*o,"Array",{find:function(t){return i(this,t,arguments.length>1?arguments[1]:void 0)}}),n(80)("find")},function(t,e,n){"use strict";n.r(e);var r=n(82),i=n.n(r);var o=n(83),a=n.n(o),s=n(84),c=n.n(s);function l(t,e){return function(t){if(i()(t))return t}(t)||function(t,e){if(c()(Object(t))||"[object Arguments]"===Object.prototype.toString.call(t)){var n=[],r=!0,i=!1,o=void 0;try{for(var s,l=a()(t);!(r=(s=l.next()).done)&&(n.push(s.value),!e||n.length!==e);r=!0);}catch(t){i=!0,o=t}finally{try{r||null==l.return||l.return()}finally{if(i)throw o}}return n}}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}n(106),n(114),n(117),n(118),n(119),n(124),n(128);var u=n(61),h=n.n(u),f=n(4),p=n.n(f);function _(t){return(_="function"==typeof p.a&&"symbol"==typeof h.a?function(t){return typeof t}:function(t){return t&&"function"==typeof p.a&&t.constructor===p.a&&t!==p.a.prototype?"symbol":typeof t})(t)}function d(t){return(d="function"==typeof p.a&&"symbol"===_(h.a)?function(t){return _(t)}:function(t){return t&&"function"==typeof p.a&&t.constructor===p.a&&t!==p.a.prototype?"symbol":_(t)})(t)}n(140),n(141),n(150),n(154),n(155),n(156),n(157);!function(t,e,n,r){n.openyMap=function(){return{locations:null,marker_image_url:null,shadow_image_url:null,map:null,tags:{},distance_limit:null,center_point:null,search_center_point:null,search_center_marker:null,geocoder:function(){return void 0!==google.maps?new google.maps.Geocoder:{}},libraryIsLoaded:function(){return void 0!==e.google},normalize_point:function(t){return{lat:t.lat(),lon:t.lng()}},toRad:function(t){return t*Math.PI/180},init:function(e){this.component_el=e.component_el,this.map_data=e.map_data,this.tags_style=e.tags_style,this.locations=this.map_data,this.marker_image_url=e.marker_image_url||null,this.shadow_image_url=e.shadow_image_url||null,this.search_center_marker=e.search_center_marker||null,this.map_el=this.component_el.find(".openy-map"),this.messages_el=this.component_el.find(".messages"),this.map_controls_el=this.component_el.find(".map_controls"),this.search_field_el=this.map_controls_el.find("input.search_field"),this.distance_limit_el=this.map_controls_el.find("select.distance_limit_value"),this.locate_me_el=this.map_controls_el.find(".locateme"),this.tags={},this.default_tags=r.openyMapSettings.default_tags,this.init_map(),this.init_tags(),this.init_map_locations(),this.draw_map_controls(),this.hookup_map_controls_events(),this.update_tag_filters(),this.update_amenities_filters(),this.draw_map_locations(),this.draw_list_locations();var n=document.location.href.match(/&?[amp;]?map_location=([\w|\+]*)&?[amp;]?/),i=this;navigator.geolocation||t(".with-geo").remove(),this.component_el.find(".zip-code .btn-submit").on("click",t.proxy(this.apply_search,this)),this.search_field_el.on("keypress",(function(t){13==t.keyCode&&i.apply_search()})),n&&(t(".search_field").val(n[1].replace(/\+/g," ")),t(".distance_limit option").eq(2).attr("selected",!0),t(".zip-code .btn-submit").click())},init_map:function(){this.map=new google.maps.Map(this.map_el[0],{scaleControl:!0,center:this.center,zoom:9,scrollwheel:!1,mapTypeId:google.maps.MapTypeId.ROADMAP}),this.init_map_center(),this.component_el.trigger("initialized",[this.map])},init_map_center:function(){this.search_center_marker=this.search_center_marker||new google.maps.Marker({position:this.center_point,animation:google.maps.Animation.DROP}),this.search_center_marker&&(this.search_center_marker.setVisible(!1),this.search_center_marker.setMap(this.map))},filter_change:function(e){if(e){var n=t(e.currentTarget),r=this.component_el.find("nav.types input."+n.attr("class")+"[type=checkbox]");r.prop("checked",n.prop("checked")),r.parents("label").toggleClass("checked",n.prop("checked"))}this.update_tag_filters(),this.update_amenities_filters(),this.redraw_map_locations(),this.draw_list_locations()},hookup_map_controls_events:function(){this.map_controls_el.find(".tag_filters input[type=checkbox]").on("change",t.proxy(this.filter_change,this)),this.search_field_el.on("change",t.proxy(this.apply_search,this)),this.distance_limit_el.on("change",t.proxy(this.apply_distance_limit,this)),this.locate_me_el.on("click",t.proxy(this.locate_me_onclick,this)),this.component_el.find("nav.types input[type=checkbox]").on("change",t.proxy(this.bar_filter_change,this)),this.search_field_el.on("autocompleteselect",t.proxy(this.apply_autocomplete_search,this)),t("#views-exposed-form-location-by-amenities-block-1").find("input[type=checkbox]").on("change",t.proxy(this.filter_change,this))},apply_search:function(){var e=this.search_field_el.val();if(""!=e){this.geocoder().geocode({address:e},t.proxy((function(t,e){if("OK"==e){if(this.search_center_point=t[0].geometry.location,t[0].geometry.bounds)this.map.fitBounds(t[0].geometry.bounds);else{var n=new google.maps.LatLngBounds;if(n.extend(this.search_center_point),n.getNorthEast().equals(n.getSouthWest())){var r=new google.maps.LatLng(n.getNorthEast().lat()+.001,n.getNorthEast().lng()+.001),i=new google.maps.LatLng(n.getNorthEast().lat()-.001,n.getNorthEast().lng()-.001);n.extend(r),n.extend(i)}this.map.fitBounds(n)}this.search_center=this.map.getCenter(),this.draw_search_center(),this.apply_distance_limit()}}),this))}else this.reset_search_results()},apply_autocomplete_search:function(e,r){var i=[];this.locations.forEach((function(t){t.name==r.item.value&&i.push(t)})),null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit="",this.search_center_marker.setPosition(this.search_center_point),this.search_center_marker.setVisible(!1);for(var o=new google.maps.LatLngBounds,a=0;a<i.length;a++){var s=i[a];o.extend(s.marker.getPosition()),s.marker.setVisible(!0)}this.map.fitBounds(o);for(var c=0;c<this.locations.length;c++)void 0!==this.locations[c].element&&(this.locations[c].element.hide(),t(this.locations[c].element).parents(".locations-list").find(".location-title").hide());if(i.length)for(var l=0;l<i.length;l++)void 0!==i[l].element&&(i[l].element.show(),t(i[l].element).parents(".locations-list").find(".location-title").show());else{var u='<div class="col-xs-12 text-center"><p>'+n.t("We're sorry no results were found in your area")+"</p></div>";this.messages_el.hide().html(u).fadeIn()}},apply_distance_limit:function(){null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit=this.distance_limit_el.val(),this.draw_search_center(),this.redraw_map_locations(),this.draw_list_locations()},reset_search_results:function(){null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit="",this.search_center_marker.setPosition(this.search_center_point),this.search_center_marker.setVisible(!1),this.redraw_map_locations(),this.draw_list_locations()},locate_me_onclick:function(e){navigator.geolocation&&(this.search_field_el.val(""),this.geolocation_watcher=navigator.geolocation.watchPosition(t.proxy(this.locate_me,this)))},locate_me:function(e){var n=e.coords.lat,r=e.coords.lng;this.search_center_point=new google.maps.LatLng(n,r),this.map.setCenter(this.search_center_point),this.map.setZoom(14),e.coords.accuracy<=15840&&(this.geocoder().geocode({latLng:this.search_center_point},t.proxy((function(t,e){t[0]&&this.search_field_el.val(t[0].formatted_address),this.apply_search()}),this)),navigator.geolocation.clearWatch(this.geolocation_watcher)),this.draw_search_center()},init_tags:function(){for(var e=0;e<this.locations.length;e++){var n=this.locations[e];n.tags||(n.tags=[]),d(n.tags)==d("")&&(n.tags=[n.tags]);for(var r=0;r<n.tags.length;r++){var i=n.tags[r];i in this.tags||(this.tags[i]={marker_icons:[]}),n.icon&&-1==t.inArray(n.icon,this.tags[i].marker_icons)&&this.tags[i].marker_icons.push(n.icon)}}},update_tag_filters:function(){this.tag_filters=[];var e=this;this.map_controls_el.find(".tag_filters input[type=checkbox]:checked").each((function(n){var r=t(this);e.tag_filters.push(r.val())}))},update_amenities_filters:function(){this.amenities_filters=[];var e=this;t("#views-exposed-form-location-by-amenities-block-1").find("input[type=checkbox]:checked").each((function(n){var r=t(this);e.amenities_filters.push(r.val())}))},apply_filters:function(t){return t=this.apply_tag_filters(t),t=this.apply_distance_filters(t),t=this.apply_amenities_filters(t),this.set_url_parameters(),t},apply_tag_filters:function(e){var n=this.tag_filters.length,r=Object.keys(this.tags).length,i=r>this.default_tags.length;if(0===n||n===r&&i)return e;for(var o=[],a=0;a<e.length;a++)for(var s=e[a],c=0;c<this.tag_filters.length;c++){var l=this.tag_filters[c];t.inArray(l,s.tags)>=0&&o.push(s)}return o},apply_distance_filters:function(t){if(!this.search_center)return t;if(!this.distance_limit||""===this.distance_limit)return t;for(var e=this.normalize_point(this.search_center),n=[],r=parseFloat(e.lat),i=parseFloat(e.lon),o=this.toRad(r),a=0;a<t.length;a++){var s=t[a],c=parseFloat(s.lat),l=parseFloat(s.lng),u=this.toRad(c-r),h=this.toRad(l-i),f=this.toRad(c),p=Math.sin(u/2)*Math.sin(u/2)+Math.sin(h/2)*Math.sin(h/2)*Math.cos(o)*Math.cos(f),_=3963*(2*Math.atan2(Math.sqrt(p),Math.sqrt(1-p)));_<=this.distance_limit&&(s.distance=_,n.push(s))}return n},apply_amenities_filters:function(e){if(0===this.amenities_filters.length)return e;for(var n=[],r=0;r<e.length;r++)for(var i=e[r],o=0;o<this.amenities_filters.length;o++){var a=this.amenities_filters[o];t.inArray(a,i.amenities)>=0&&n.push(i)}return n},init_active_tags:function(){if(this.initial_active_tags)return this.initial_active_tags;var e=[],n=this.get_parameters().type,r=n?n.split(","):[];for(var i in this.tags)0===r.length?e.push(i):t.inArray(this.encode_to_url_format(i),r)>=0&&e.push(i);this.initial_active_tags=e,0===r.length&&(this.initial_active_tags=this.default_tags)},get_parameters:function(){for(var t=e.location.search.substring(1).split("&"),n={},r=0;r<t.length;r++){var i=t[r].split("=");n[unescape(i[0])]=unescape(i[1])}return n},set_url_parameters:function(){var n=document.location.pathname,r=this.get_parameters(),i=this.tag_filters,o=this.amenities_filters,a="",s="",c=t(".search_field").val()||r.hasOwnProperty("map_location")&&r.map_location||"";c&&(c="?map_location="+this.encode_to_url_format(c)),i&&(a=c?"&":"?",a+="type=",i.forEach((function(t){a+=this.encode_to_url_format(t)+","}),this,a),a=a.substring(0,a.length-1)),o&&(s="&",s+="amenities=",o.forEach((function(t){s+=this.encode_to_url_format(t)+","}),this,s),s=s.substring(0,s.length-1)),e.history.replaceState(null,null,n+c+a+s)},draw_map_controls:function(){this.init_active_tags();var e="";if("list-checkboxes"==this.tags_style){e='<select id="tag-filter" class="form-control multiselect" name="tag_filter" multiple="multiple">';var n=Object.keys(this.tags).sort();n.splice(n.indexOf("YMCA"),1),n.splice(n.indexOf("Camps"),1),n.unshift("Camps"),n.unshift("YMCA"),n.forEach((function(n){var r="";t.inArray(n,this.initial_active_tags)>=0&&(r="selected"),e+='<option value="'+n+'" '+r+">"+n+"</option>"}),this),e+="</select>"}else for(var r in this.tags){var i="";t.inArray(r,this.initial_active_tags)>=0&&(i='checked="checked"');var o='<label class="btn btn-default" for="tag_'+r+'">';o+='<input autocomplete="off" id="tag_'+r+'" class="tag_'+r+'" type="checkbox" value="'+r+'" '+i+"/>"+r;for(var a=0;a<this.tags[r].marker_icons.length;a++)o+='<img class="tag_icon inline-hidden-sm" src="'+this.tags[r].marker_icons[a]+'" aria-hidden="true" />';e+=o+="</label>"}this.map_controls_el.find(".tag_filters").append(e),"list-checkboxes"==this.tags_style&&t(".tag_filters .multiselect").multiselect({columns:1,showCheckbox:!0,minHeight:50,texts:{placeholder:"Select options"},onOptionClick:t.proxy(this.filter_change,this)});var s=[];this.locations.forEach((function(t){s.push(t.name)})),this.search_field_el.autocomplete({minLength:3,source:s})},encode_to_url_format:function(t){return t.toLowerCase().replace(/[^\w ]+/g,"").replace(/ +/g,"-")},draw_map_locations:function(){var t=this.apply_filters(this.locations);if(0!==t.length){for(var e=new google.maps.LatLngBounds,n=0;n<t.length;n++){var r=t[n];e.extend(r.marker.getPosition()),r.marker.setVisible(!0)}if(e.getNorthEast().equals(e.getSouthWest())){var i=new google.maps.LatLng(e.getNorthEast().lat()+.001,e.getNorthEast().lng()+.001),o=new google.maps.LatLng(e.getNorthEast().lat()-.001,e.getNorthEast().lng()-.001);e.extend(i),e.extend(o)}this.map.fitBounds(e)}else this.map.setCenter(this.search_center_point)},redraw_map_locations:function(){for(var t=0;t<this.locations.length;t++){this.locations[t].marker.setVisible(!1)}this.draw_map_locations()},draw_list_locations:function(){for(var e=this.apply_filters(this.locations),r=0;r<this.locations.length;r++)void 0!==this.locations[r].element&&(this.locations[r].element.hide(),t(this.locations[r].element).parents(".locations-list").find(".location-title").hide());if(e.length){this.messages_el.hide();for(var i=0;i<e.length;i++)void 0!==e[i].element&&(e[i].element.show(),t(e[i].element).parents(".locations-list").find(".location-title").show())}else{var o='<div class="col-xs-12 text-center"><p>'+n.t("No locations were found in this area. Please try a different area or increase your search distance.")+"</p></div>";this.messages_el.hide().html(o).fadeIn()}},draw_list_location:function(t){return t.markup},draw_map_location:function(t){return this.draw_list_location(t)},init_map_locations:function(){for(var t=this,e=this.locations,n=function(e,n){return function(){for(var r=0;r<t.locations.length;r++)t.locations[r].infowindow.close();e.open(this.map,n)}},r=0;r<e.length;r++){var i=e[r];i.point=new google.maps.LatLng(i.lat,i.lng);var o='<div class="marker_tooltip">'+this.draw_map_location(i)+"</div>",a=new google.maps.MarkerImage(this.marker_image_url)||null;a=i.icon?new google.maps.MarkerImage(i.icon):a;var s=i.shadow?new google.maps.MarkerImage(i.shadow):null,c=new google.maps.Marker({position:i.point,icon:a,shadow:s,animation:google.maps.Animation.DROP}),l=new google.maps.InfoWindow({content:o});i.infowindow=l,google.maps.event.addListener(c,"click",n(l,c)),c.setVisible(!1),c.setMap(this.map),i.marker=c}},draw_search_center:function(){this.search_center_marker.setPosition(this.search_center_point),this.search_center_marker.setVisible(!0)},bar_filter_change:function(e){var n=t(e.currentTarget),r=this.map_controls_el.find("input."+n.attr("class")+"[type=checkbox]");this.map_controls_el.find("input[type=checkbox]").prop("checked",!1),this.component_el.find("nav.types input[type=checkbox]").prop("checked",!1),this.component_el.find("nav.types label").removeClass("checked"),r.prop("checked",!0),n.prop("checked",!0),n.parents("label").addClass("checked"),this.filter_change()},build_google_url:function(t){return(t=(t=t.trim()).replace(/ /g,"+")).length>0&&(t+="+"),t}}},n.baseLayerWikimedia={tilePattern:"https://maps.wikimedia.org/osm-intl/{z}/{x}/{y}{r}.png",options:{attribution:'<a href="https://wikimediafoundation.org/wiki/Maps_Terms_of_Use">Wikimedia</a>',minZoom:1,maxZoom:19}},n.baseLayerEsriWorldStreetMap={tilePattern:"https://server.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer/tile/{z}/{y}/{x}",options:{attribution:"Tiles &copy; Esri &mdash; Source: Esri, DeLorme, NAVTEQ, USGS, Intermap, iPC, NRCAN, Esri Japan, METI, Esri China (Hong Kong), Esri (Thailand), TomTom, 2012"}},n.baseLayerEsriNatGeoWorldMap={tilePattern:"https://server.arcgisonline.com/ArcGIS/rest/services/NatGeo_World_Map/MapServer/tile/{z}/{y}/{x}",options:{attribution:"Tiles &copy; Esri &mdash; National Geographic, Esri, DeLorme, NAVTEQ, UNEP-WCMC, USGS, NASA, ESA, METI, NRCAN, GEBCO, NOAA, iPC",maxZoom:16}},n.baseLayerOpenStreetMapMapnik={tilePattern:"https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",options:{attribution:'&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>',maxZoom:19}},n.openyMapLeaflet=function(){return{baseLayer:n.baseLayerWikimedia,locations:null,marker_image_url:null,shadow_image_url:null,map:null,tags:{},distance_limit:null,center_point:null,search_center_point:null,search_center_marker:null,fitBoundsOptions:null,default_search_location:null,libraryIsLoaded:function(){return void 0!==e.L},normalize_point:function(t){return{lat:t.lat(),lon:t.lng()}},toRad:function(t){return t*Math.PI/180},init:function(e){this.component_el=e.component_el,this.map_data=e.map_data,this.tags_style=e.tags_style,this.locations=this.map_data,this.fitBoundsOptions={paddingTopLeft:L.point(0,40),paddingBottomRight:L.point(0,10)},this.marker_image_url=e.marker_image_url||null,this.shadow_image_url=e.shadow_image_url||null,this.search_center_marker=e.search_center_marker||null,this.map_el=this.component_el.find(".openy-map"),this.messages_el=this.component_el.find(".messages"),this.map_controls_el=this.component_el.find(".map_controls"),this.search_field_el=this.map_controls_el.find("input.search_field"),this.distance_limit_el=this.map_controls_el.find("select.distance_limit_value"),this.locate_me_el=this.map_controls_el.find(".locateme"),this.tags={},this.default_tags=r.openyMapSettings.default_tags,this.init_map(),this.init_tags(),this.init_map_locations(),this.draw_map_controls(),this.hookup_map_controls_events(),this.update_tag_filters(),this.update_amenities_filters(),this.draw_map_locations(),this.draw_list_locations();var n=document.location.href.match(/&?[amp;]?map_location=([\w|\+]*)&?[amp;]?/),i=this;navigator.geolocation||t(".with-geo").remove(),this.component_el.find(".zip-code .btn-submit").on("click",t.proxy(this.apply_search,this)),this.search_field_el.on("keypress",(function(t){13==t.keyCode&&i.apply_search()})),n&&(t(".search_field").val(n[1].replace(/\+/g," ")),t(".distance_limit option").eq(2).attr("selected",!0),t(".zip-code .btn-submit").click())},init_map:function(){this.map=L.map(this.map_el[0]).setView([51.505,-.09],13),L.tileLayer(this.baseLayer.tilePattern,this.baseLayer.options).addTo(this.map),this.map.scrollWheelZoom.disable(),this.init_map_center()},init_map_center:function(){var t=L.icon({iconUrl:this.search_icon,iconRetinaUrl:this.search_icon_retina,shadowUrl:"https://cdnjs.cloudflare.com/ajax/libs/leaflet/0.7.7/images/marker-shadow.png",iconSize:[25,41],iconAnchor:[12,41],popupAnchor:[1,-34],shadowSize:[41,41]});this.search_center_marker=this.search_center_marker||L.marker(this.map.getCenter(),{icon:t}),this.search_center_marker&&this.search_center_marker.removeFrom(this.map)},filter_change:function(e){if(e){var n=t(e.currentTarget),r=this.component_el.find("nav.types input."+n.attr("class")+"[type=checkbox]");r.prop("checked",n.prop("checked")),r.parents("label").toggleClass("checked",n.prop("checked"))}this.update_tag_filters(),this.update_amenities_filters(),this.redraw_map_locations(),this.draw_list_locations()},hookup_map_controls_events:function(){this.map_controls_el.find(".tag_filters input[type=checkbox]").on("change",t.proxy(this.filter_change,this)),this.search_field_el.on("change",t.proxy(this.apply_search,this)),this.distance_limit_el.on("change",t.proxy(this.apply_distance_limit,this)),this.locate_me_el.on("click",t.proxy(this.locate_me_onclick,this)),this.component_el.find("nav.types input[type=checkbox]").on("change",t.proxy(this.bar_filter_change,this)),this.search_field_el.on("autocompleteselect",t.proxy(this.apply_autocomplete_search,this)),t("#views-exposed-form-location-by-amenities-block-1").find("input[type=checkbox]").on("change",t.proxy(this.filter_change,this))},apply_search:function(){var t=this,e=this.search_field_el.val();""!=e?this.geocode(e,(function(e,n){if("success"==n&&e.length>0){if(t.search_center_point=L.latLng(e[0].lat,e[0].lon),e[0].boundingbox){var r=L.latLngBounds();r.extend(L.latLng(e[0].boundingbox[0],e[0].boundingbox[2])),r.extend(L.latLng(e[0].boundingbox[1],e[0].boundingbox[3])),t.map.fitBounds(r,t.fitBoundsOptions)}t.search_center=t.search_center_point,t.draw_search_center(),t.apply_distance_limit()}})):this.reset_search_results()},geocode:function(e,n){var r=this.default_search_location?"+"+this.default_search_location:"";t.getJSON("https://nominatim.openstreetmap.org/search?format=json&limit=5&q="+e+r,n)},apply_autocomplete_search:function(e,r){var i=[];this.locations.forEach((function(t){t.name==r.item.value&&i.push(t)})),null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit="",this.search_center_marker.removeFrom(this.maps);for(var o=L.latLngBounds(),a=0;a<i.length;a++){var s=i[a];o.extend(s.point),s.marker.addTo(this.map)}this.map.fitBounds(o,this.fitBoundsOptions);for(var c=0;c<this.locations.length;c++)void 0!==this.locations[c].element&&(this.locations[c].element.hide(),t(this.locations[c].element).parents(".locations-list").find(".location-title").hide());if(i.length)for(var l=0;l<i.length;l++)void 0!==i[l].element&&(i[l].element.show(),t(i[l].element).parents(".locations-list").find(".location-title").show());else{var u='<div class="col-xs-12 text-center"><p>'+n.t("We're sorry no results were found in your area")+"</p></div>";this.messages_el.hide().html(u).fadeIn()}},apply_distance_limit:function(){null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit=this.distance_limit_el.val(),this.draw_search_center(),this.redraw_map_locations(),this.draw_list_locations()},reset_search_results:function(){null===this.search_center&&(this.search_center=this.map.getCenter()),this.distance_limit="",this.search_center_point&&(this.search_center_marker.setLatLng(this.search_center_point),this.search_center_marker.addTo(this.map)),this.redraw_map_locations(),this.draw_list_locations(),this.set_url_parameters()},locate_me_onclick:function(e){navigator.geolocation&&(this.search_field_el.val(""),this.geolocation_watcher=navigator.geolocation.watchPosition(t.proxy(this.locate_me,this)))},locate_me:function(t){var e=t.coords.lat,n=t.coords.lng;this.search_center_point=L.latLng(e,n),this.map.setView(this.search_center_point,14),t.coords.accuracy<=15840&&navigator.geolocation.clearWatch(this.geolocation_watcher),this.draw_search_center()},init_tags:function(){for(var e=0;e<this.locations.length;e++){var n=this.locations[e];n.tags||(n.tags=[]),d(n.tags)==d("")&&(n.tags=[n.tags]);for(var r=0;r<n.tags.length;r++){var i=n.tags[r];i in this.tags||(this.tags[i]={marker_icons:[]}),n.icon&&-1==t.inArray(n.icon,this.tags[i].marker_icons)&&this.tags[i].marker_icons.push(n.icon)}}},update_tag_filters:function(){this.tag_filters=[];var e=this;this.map_controls_el.find(".tag_filters input[type=checkbox]:checked").each((function(n){var r=t(this);e.tag_filters.push(r.val())}))},update_amenities_filters:function(){this.amenities_filters=[];var e=this;t("#views-exposed-form-location-by-amenities-block-1").find("input[type=checkbox]:checked").each((function(n){var r=t(this);e.amenities_filters.push(r.val())}))},apply_filters:function(t){return t=this.apply_tag_filters(t),t=this.apply_distance_filters(t),t=this.apply_amenities_filters(t),this.set_url_parameters(),t},apply_tag_filters:function(e){var n=this.tag_filters.length,r=Object.keys(this.tags).length,i=r>this.default_tags.length;if(0===n||n===r&&i)return e;for(var o=[],a=0;a<e.length;a++)for(var s=e[a],c=0;c<this.tag_filters.length;c++){var l=this.tag_filters[c];t.inArray(l,s.tags)>=0&&o.push(s)}return o},apply_distance_filters:function(t){if(!this.search_center)return t;if(!this.distance_limit||""===this.distance_limit)return t;for(var e=this.search_center,n=[],r=parseFloat(e.lat),i=parseFloat(e.lng),o=this.toRad(r),a=0;a<t.length;a++){var s=t[a],c=parseFloat(s.point.lat),l=parseFloat(s.point.lng),u=this.toRad(c-r),h=this.toRad(l-i),f=this.toRad(c),p=Math.sin(u/2)*Math.sin(u/2)+Math.sin(h/2)*Math.sin(h/2)*Math.cos(o)*Math.cos(f),_=3963*(2*Math.atan2(Math.sqrt(p),Math.sqrt(1-p)));_<=this.distance_limit&&(s.distance=_,n.push(s))}return n},apply_amenities_filters:function(e){if(0===this.amenities_filters.length)return e;for(var n=[],r=0;r<e.length;r++)for(var i=e[r],o=0;o<this.amenities_filters.length;o++){var a=this.amenities_filters[o];t.inArray(a,i.amenities)>=0&&n.push(i)}return n},init_active_tags:function(){if(this.initial_active_tags)return this.initial_active_tags;var e=[],n=this.get_parameters().type,r=n?n.split(","):[];for(var i in this.tags)0===r.length?e.push(i):t.inArray(this.encode_to_url_format(i),r)>=0&&e.push(i);this.initial_active_tags=e,0===r.length&&(this.initial_active_tags=this.default_tags)},get_parameters:function(){for(var t=e.location.search.substring(1).split("&"),n=[],r=0;r<t.length;r++){var i=t[r].split("=");n[unescape(i[0])]=unescape(i[1])}return n},set_url_parameters:function(){for(var n=document.location.pathname,r=this.get_parameters(),i=this.tag_filters,o=this.amenities_filters,a="",s="",c={},u="",h=t(".search_field").val()||r.hasOwnProperty("map_location")&&r.map_location||"",f=0,p=Object.entries(r);f<p.length;f++){var _=l(p[f],2),d=_[0],g=_[1];"utm"===d.split("_")[0]&&(c[d]=g)}Object.keys(c).length&&(u="&",u+=jQuery.param(c)),h&&(h="?map_location="+this.encode_to_url_format(h)),i&&(a=h?"&":"?",a+="type=",i.forEach((function(t){a+=this.encode_to_url_format(t)+","}),this,a),a=a.substring(0,a.length-1)),o&&(s="&",s+="amenities=",o.forEach((function(t){s+=this.encode_to_url_format(t)+","}),this,s),s=s.substring(0,s.length-1)),e.history.replaceState(null,null,n+h+a+s+u)},draw_map_controls:function(){this.init_active_tags();var e="";if("list-checkboxes"==this.tags_style){e='<select id="tag-filter" class="form-control multiselect" name="tag_filter" multiple="multiple">';var n=Object.keys(this.tags).sort();n.splice(n.indexOf("YMCA"),1),n.splice(n.indexOf("Camps"),1),n.unshift("Camps"),n.unshift("YMCA"),n.forEach((function(n){var r="";t.inArray(n,this.initial_active_tags)>=0&&(r="selected"),e+='<option value="'+n+'" '+r+">"+n+"</option>"}),this),e+="</select>"}else for(var r in this.tags){var i="";t.inArray(r,this.initial_active_tags)>=0&&(i='checked="checked"');var o='<label class="btn btn-default" for="tag_'+r+'">';o+='<input autocomplete="off" id="tag_'+r+'" class="tag_'+r+'" type="checkbox" value="'+r+'" '+i+"/>"+r;for(var a=0;a<this.tags[r].marker_icons.length;a++)o+='<img class="tag_icon inline-hidden-sm" src="'+this.tags[r].marker_icons[a]+'" aria-hidden="true" />';e+=o+="</label>"}this.map_controls_el.find(".tag_filters").append(e),"list-checkboxes"==this.tags_style&&t(".tag_filters .multiselect").multiselect({columns:1,showCheckbox:!0,minHeight:50,texts:{placeholder:"Select options"},onOptionClick:t.proxy(this.filter_change,this)});var s=[];this.locations.forEach((function(t){s.push(t.name)})),this.search_field_el.autocomplete({minLength:3,source:s})},encode_to_url_format:function(t){return t.toLowerCase().replace(/[^\w ]+/g,"").replace(/ +/g,"-")},draw_map_locations:function(){var t=this.apply_filters(this.locations);if(0!==t.length){for(var e=L.latLngBounds([]),n=0;n<t.length;n++){var r=t[n];e.extend(r.point),r.marker.addTo(this.map)}if(e.getNorthEast().equals(e.getSouthWest())){var i=L.latLng(e.getNorthEast().lat+.001,e.getNorthEast().lng+.001),o=L.latLng(e.getNorthEast().lat-.001,e.getNorthEast().lng-.001);e.extend(i),e.extend(o)}this.map.fitBounds(e,this.fitBoundsOptions)}else null!==this.search_center_point&&this.map.setView(this.search_center_point)},redraw_map_locations:function(){for(var t=0;t<this.locations.length;t++){this.locations[t].marker.removeFrom(this.map)}this.draw_map_locations()},draw_list_locations:function(){for(var e=this.apply_filters(this.locations),r=0;r<this.locations.length;r++)void 0!==this.locations[r].element&&(this.locations[r].element.hide(),t(this.locations[r].element).parents(".locations-list").find(".location-title").hide());if(e.length){this.messages_el.hide();for(var i=0;i<e.length;i++)void 0!==e[i].element&&(e[i].element.show(),t(e[i].element).parents(".locations-list").find(".location-title").show())}else{var o='<div class="col-xs-12 text-center"><p>'+n.t("No locations were found in this area. Please try a different area or increase your search distance.")+"</p></div>";this.messages_el.hide().html(o).fadeIn()}},draw_list_location:function(t){return t.markup},draw_map_location:function(t){return this.draw_list_location(t)},init_map_locations:function(){for(var e=this.locations,n=["iconSize","shadowSize","iconAnchor","shadowAnchor","popupAnchor"],r=0;r<e.length;r++){var i=e[r];i.point=L.latLng(i.lat,i.lng);var o='<div class="marker_tooltip">'+this.draw_map_location(i)+"</div>",a={iconUrl:i.icon,iconSize:[32,42],iconAnchor:[16,38],popupAnchor:[0,-36]};t(n).each((function(t){void 0!==i[t]&&(a[t]=i[t])}));var s=i.icon?L.icon(a):new L.Icon.Default,c=L.marker(i.point,{icon:s});c.bindPopup(o,{maxWidth:180}).openPopup(),i.marker=c}},draw_search_center:function(){this.search_center_point&&(this.search_center_marker.setLatLng(this.search_center_point),this.search_center_marker.addTo(this.map))},bar_filter_change:function(e){var n=t(e.currentTarget),r=this.map_controls_el.find("input."+n.attr("class")+"[type=checkbox]");this.map_controls_el.find("input[type=checkbox]").prop("checked",!1),this.component_el.find("nav.types input[type=checkbox]").prop("checked",!1),this.component_el.find("nav.types label").removeClass("checked"),r.prop("checked",!0),n.prop("checked",!0),n.parents("label").addClass("checked"),this.filter_change()}}},n.behaviors.openyMap={attach:function(e,r){if(void 0!==r.openyMap&&void 0!==r.openyMapSettings){var i,o=r.openyMap;switch(r.openyMapSettings.engine){case"gmaps":i=new n.openyMap;break;case"leaflet":default:switch((i=new n.openyMapLeaflet).default_search_location=r.openyMapSettings.default_location,i.search_icon=r.openyMapSettings.search_icon,i.search_icon_retina=r.openyMapSettings.search_icon_retina,r.openyMapSettings.base_layer){case"Esri.WorldStreetMap":i.baseLayer=n.baseLayerEsriWorldStreetMap;break;case"Esri.NatGeoWorldMap":i.baseLayer=n.baseLayerEsriNatGeoWorldMap;break;case"OpenStreetMap.Mapnik":i.baseLayer=n.baseLayerOpenStreetMapMapnik;break;case"Wikimedia":i.baseLayer=n.baseLayerWikimedia}}t(".locations-list .node--view-mode-teaser").each((function(){for(var e=t(this),n=0;n<o.length;n++)void 0!==o[n]&&"undefined"!==e.find(".location-item--title")[0].innerText&&t.trim(e.find(".location-item--title")[0].innerText).toLowerCase()==t.trim(o[n].name).toLocaleLowerCase()&&(o[n].element={},o[n].element=e.parent(),o[n].amenities=[],o[n].amenities=e.data("amenities"))})),t(".openy-map-canvas",e).once().each((function(){var e=t(this),n=setInterval((function(){i.libraryIsLoaded()&&(i.init({component_el:e.closest(".openy-map-wrapper"),map_data:o,tags_style:e.closest(".location-finder-filters").attr("data-tags-style")}),r.openyMap=[],clearInterval(n))}),100)}))}}}}(jQuery,window,Drupal,drupalSettings)}]);
