+++ modules/custom/openy_map/js/map/babel.config.js
@@ -1,0 +1,14 @@
+// babel.config.js
+module.exports = {
+  presets: [
+    ['@vue/app', {
+      polyfills: [
+        'es6.promise',
+        'es6.symbol',
+        'es6.array.iterator',
+        'es6.object.assign',
+        'es7.promise.finally'
+      ]
+    }]
+  ]
+}
